% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/Read_data.R
\name{Add10XMetrics}
\alias{Add10XMetrics}
\alias{Add10XMetrics.Seurat}
\alias{Add10XMetrics.default}
\title{Add 10X Metrics to Object}
\usage{
Add10XMetrics(object, Metrics, ...)

\method{Add10XMetrics}{Seurat}(object, Metrics, ...)

\method{Add10XMetrics}{default}(object, Metrics, ...)
}
\arguments{
\item{object}{An object}

\item{Metrics}{10X Metrics generated by \code{Read10XMetrics} function}

\item{...}{Additional arguments passed to methods}
}
\value{
Object with 10X Metrics added
}
\description{
Add 10X Metrics to an object
}
