% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFeature.R
\name{AddFeaturePlotLabel}
\alias{AddFeaturePlotLabel}
\title{Add feature plot labels}
\usage{
AddFeaturePlotLabel(p, metric_order_col, ntop = 10)
}
\arguments{
\item{p}{A ggplot object. It's expected that the data within the ggplot
object contains a 'Feature' column.}

\item{metric_order_col}{A string, the name of the column in the plot data
to be used for sorting features (e.g., 'variance'). Features with
higher values in this column will be prioritized.}

\item{ntop}{An integer, the number of top features to label. Defaults to 10.}
}
\value{
A modified ggplot object with added text labels for the top
'ntop' features.
}
\description{
This function takes a ggplot object, sorts its data by a specified metric
column in descending order, and then adds labels for the top 'ntop'
features using \code{ggrepel::geom_text_repel}. This helps highlight
the most prominent features on the plot.
}
