% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMQCData.R
\name{AddSingleCellMQCData}
\alias{AddSingleCellMQCData}
\title{Add Single Cell Multi-omics QC Data to a Seurat Object}
\usage{
AddSingleCellMQCData(object, listname, metadata, sample.by = "orig.ident")
}
\arguments{
\item{object}{A Seurat object.}

\item{listname}{The name of the list within the \code{SingleCellMQC} slot of the \code{misc} object where the metadata should be stored.}

\item{metadata}{A data frame containing the new metadata to add. Each row corresponds to a cell, and columns represent different metadata attributes.The column name in \code{metadata} used to match rows with cells in the Seurat object.}

\item{sample.by}{Defaults to "orig.ident".}
}
\value{
The Seurat object with the updated \code{SingleCellMQC} data in its \code{misc}.
}
\description{
This function adds or updates metadata in the Seurat object's miscellaneous (\code{misc}) slot, specifically targeting the Single Cell MQC (Quality Control) data.
}
\seealso{
\code{\link{GetSingleCellMQCData}} for retrieving Single Cell MQC data from a Seurat object.
}
