% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_effect.R
\name{AggregateGroupAcrossCells}
\alias{AggregateGroupAcrossCells}
\title{Aggregate Counts by Group Across Cells}
\usage{
AggregateGroupAcrossCells(object, metadata, group.by)
}
\arguments{
\item{object}{A matrix or data frame containing the count data, where rows represent features (e.g., genes or proteins) and columns represent cells.}

\item{metadata}{A data frame containing metadata information for the cells. The number of rows should match the number of columns in \code{object}.}

\item{group.by}{A character string specifying the column in \code{metadata} to group cells by.}
}
\value{
A matrix where rows are features and columns are groups, with values representing the summed counts for each group.
}
\description{
This function aggregates count data from a given matrix or data frame by grouping cells based on a specified metadata column. The output is a matrix where rows represent features and columns represent groups, with the values being the summed counts of the corresponding group.
}
\examples{

\dontrun{
# Aggregate counts by group
counts <- matrix(rpois(1000, lambda = 10), nrow = 100, ncol = 10)
metadata <- data.frame(group = rep(c("A", "B"), each = 5))
grouped_counts <- AggregateGroupAcrossCells(counts, metadata, group.by = "group")
}

}
