% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetricsPerFeature}
\alias{CalculateMetricsPerFeature}
\title{Calculate Feature QC Metrics}
\usage{
CalculateMetricsPerFeature(
  object,
  add.Seurat = T,
  verbose = TRUE,
  sample.by = "orig.ident"
)
}
\arguments{
\item{object}{A Seurat object.}

\item{add.Seurat}{Logical, if TRUE, adds the computed metrics to the Seurat object. Otherwise, returns a list or data.frame with the metrics.}

\item{verbose}{A logical value. If \code{TRUE} (default), the message will be printed.
If \code{FALSE} , the message will be suppressed (not printed).}

\item{sample.by}{The column name by which to group samples for metrics computation.}
}
\value{
The Seurat object with feature metrics added if \code{add.Seurat} is TRUE, otherwise a list or data frame of feature qc metrics.
}
\description{
Calculate metrics at the feature level, such as variable features and expression statistics.
}
\examples{
\dontrun{
data <- CalculateMetricsPerFeature(data, add.Seurat= T)
}
}
