% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetricsPerSample.count}
\alias{CalculateMetricsPerSample.count}
\title{Calculate Count Metrics for Samples}
\usage{
CalculateMetricsPerSample.count(object, sample.by = "orig.ident")
}
\arguments{
\item{object}{A Seurat object.}

\item{sample.by}{The column name by which to group samples for metrics computation.}
}
\value{
A data frame with samples as rows and count types as columns.
}
\description{
Calculate such as nCell, nGene/Protein for each sample in a Seurat object.
}
\examples{
\dontrun{
# Assuming `seuratObj` is a Seurat object
seuratObj <- CalculateMetricsPerSample.count(seuratObj, sample.by = "orig.ident")
}
}
\seealso{
\code{\link{CalculateMetricsPerSample}}, \code{\link{CalculateMetricsPerSample.summary}}
}
