% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetricsPerSample}
\alias{CalculateMetricsPerSample}
\title{Calculate Sample QC Metrics}
\usage{
CalculateMetricsPerSample(
  object,
  sample.by = "orig.ident",
  add.Seurat = T,
  metrics = c("nCount_RNA", "nFeature_RNA", "nCount_ADT", "nFeature_ADT", "percent.mt",
    "percent.rb", "percent.hb", "percent.dissociation", "per_feature_count_RNA",
    "per_feature_count_ADT")
)
}
\arguments{
\item{object}{A Seurat object.}

\item{sample.by}{The column name by which to group samples for metrics computation.}

\item{add.Seurat}{Logical indicating whether to add metrics directly to a Seurat object. If TRUE, adds the computed metrics to the Seurat object's metadata and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a list with the metrics.}

\item{metrics}{A vector of qc metric names to be computed. Default: "nCount_RNA", "nFeature_RNA", "nCount_ADT", "nFeature_ADT", "percent.mt", "percent.rb", "percent.hb", "percent.dissociation", "per_feature_count_RNA", "per_feature_count_ADT".}
}
\value{
The Seurat object with sample metrics added, or a list of metrics if \code{add.Seurat} is FALSE.
}
\description{
Calculate sample qc metrics, including nCell, nGene/Protein, nChain, etc.
}
\examples{
\dontrun{
data <- CalculateMetricsPerSample(data, sample.by = "orig.ident", add.Seurat = T)
}
}
\seealso{
\code{\link{CalculateMetricsPerSample.count}}, \code{\link{CalculateMetricsPerSample.summary}}
}
