% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetricsPerSample.summary}
\alias{CalculateMetricsPerSample.summary}
\title{Calculate Summary Metrics for Samples}
\usage{
CalculateMetricsPerSample.summary(
  object,
  sample.by = "orig.ident",
  metrics = c("nCount_RNA", "nFeature_RNA", "nCount_ADT", "nFeature_ADT", "percent.mt",
    "percent.rb", "percent.hb", "percent.dissociation", "per_feature_count_RNA",
    "per_feature_count_ADT")
)
}
\arguments{
\item{object}{A Seurat object or a metadata list to compute the summaries on.}

\item{sample.by}{The column name by which to group samples for metrics computation.}

\item{metrics}{A character vector of metric names for which to calculate summaries. Default:"nCount_RNA","nFeature_RNA","nCount_ADT","nFeature_ADT","percent.mt","percent.rb","percent.hb","percent.dissociation", "per_feature_count_RNA", "per_feature_count_ADT".}
}
\value{
A list of data frames, each containing summary statistics for a metric across samples.
}
\description{
Calculate summary statistics for specified metrics for each sample in a Seurat object or a metadata list.
}
\seealso{
\code{\link{CalculateMetricsPerSample}}, \code{\link{CalculateMetricsPerSample.count}}
}
