% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{CellRangerAlerts}
\alias{CellRangerAlerts}
\title{Generate Cell Ranger Alerts for QC Metrics}
\usage{
CellRangerAlerts(object, return.type = "table")
}
\arguments{
\item{object}{The input object, either a Seurat object or a data frame containing QC metrics.}

\item{return.type}{A character string indicating the return type: either "table" or "interactive_table".
Default is "table". Accepted values are "table" or "interactive_table".}
}
\value{
Depending on the \code{return.type} parameter, this function returns:
- A data frame if \code{return.type} is "table".
- An interactive HTML table if \code{return.type} is "interactive_table".
}
\description{
This function generates alerts for 10X Genomics Single Cell QC metrics.
}
\examples{
\dontrun{
# Example 1: Using a Seurat object
alerts <- CellRangerAlerts(seurat_object, return.type = "table")

# Example 2: Using a data frame directly
alerts <- CellRangerAlerts(qc_metrics_df, return.type = "interactive_table")
}

}
