% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R, R/untils.R
\name{ConvertToBPCells}
\alias{ConvertToBPCells}
\alias{ConvertToBPCells.Seurat}
\alias{ConvertToBPCells.dgCMatrix}
\alias{ConvertToBPCells.IterableMatrix}
\title{Convert Data to BPCells Format}
\usage{
ConvertToBPCells(object, ...)

\method{ConvertToBPCells}{Seurat}(
  object,
  BPdir = "./BPCellData/",
  assay = "RNA",
  slot = "counts",
  ...
)

\method{ConvertToBPCells}{dgCMatrix}(object, BPdir = "./BPCellData/", ...)

\method{ConvertToBPCells}{IterableMatrix}(object, BPdir = "./BPCellData/", ...)
}
\arguments{
\item{object}{An object to be converted}

\item{...}{Additional arguments passed to methods}

\item{BPdir}{Directory path to store BPCells data. Default is \verb{./BPCellData/}}

\item{assay}{Name of assay to use (for Seurat objects). Default is \code{RNA}}

\item{slot}{Name of slot to use (for Seurat objects). Default is \code{counts}}
}
\value{
\strong{For Seurat object}:
Returns a Seurat object with the specified assay and slot updated to BPCells format.

\strong{For dgCMatrix object}:
Returns a BPCells matrix object.
}
\description{
Convert Seurat or dgCMatrix objects to BPCells format for efficient storage and manipulation
}
\examples{
\dontrun{
# For Seurat object
seurat_obj <- ConvertToBPCells(seurat_obj)

# For dgCMatrix
obj <- ConvertToBPCells(dgcmatrix_obj)
}

}
