% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{FilterCells}
\alias{FilterCells}
\title{Filter Cells in a Seurat Object Based on Multiple Criteria}
\usage{
FilterCells(
  object,
  filter_columns = NULL,
  filter_logic = "or",
  split.by = "orig.ident",
  return.table = FALSE
)
}
\arguments{
\item{object}{A Seurat object to be filtered.}

\item{filter_columns}{A character vector of column names in the Seurat object's metadata
that contain "Pass" or "Fail" values.}

\item{filter_logic}{A string specifying the filtering logic. Must be either "and" or "or".  Default is "or".
- \code{and}: Filter cells where all specified columns are "Fail".
- \code{or}: Filter cells where at least one of the specified columns is "Fail".}

\item{split.by}{(Optional) A column name in the Seurat object's metadata to split the data by.
The function will report the number of cells before and after filtering for each group.
Default is "orig.ident".}

\item{return.table}{Logical. If TRUE, returns a data.frame summarizing the number of cells
before and after filtering, including cases with and without split.by.
If FALSE, returns the filtered Seurat object. Default is FALSE.}
}
\value{
If \code{return.table} is TRUE, returns a data.frame summarizing the number of cells
before and after filtering. If \code{return.table} is FALSE, returns the filtered Seurat object.
}
\description{
This function filters cells in a Seurat object based on multiple criteria columns.
Each column in the Seurat object's metadata should contain "Pass" or "Fail" values,
and optionally NA values. The function can filter cells based on two logical conditions:
\enumerate{
\item "and": Filter cells where all specified columns are "Fail".
\item "or": Filter cells where at least one of the specified columns is "Fail".
}
}
