% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{FindAllMarkerCOSG}
\alias{FindAllMarkerCOSG}
\title{Find markers by COSG Method}
\usage{
FindAllMarkerCOSG(
  object,
  cluster.by = "seurat_clusters",
  top_gene = 50,
  assay = "RNA"
)
}
\arguments{
\item{object}{A Seurat object. The input must be a valid Seurat object.}

\item{cluster.by}{Character. The metadata column used for clustering the cells. Default is "seurat_clusters".}

\item{top_gene}{Numeric. The number of top marker genes to identify for each cluster. Default is 50.}

\item{assay}{Character. The name of the assay to use. Default is "RNA".}
}
\value{
A data frame containing the marker genes, cluster information, COSG scores, log fold changes, mean expressions, and percentages.
}
\description{
This function identifies cluster-specific marker genes using the COSG package. It also calculates log fold changes,
the percentage of cells expressing each gene, and other related statistics for each cluster.
}
