% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_QC.R
\name{FindContaminationFeature}
\alias{FindContaminationFeature}
\title{Detect the comtamination causing genes by scCDC package}
\usage{
FindContaminationFeature(
  object,
  assay = "RNA",
  group.by = "seurat_clusters",
  split.by = "orig.ident",
  restriction_factor = 0.5,
  ...
)
}
\arguments{
\item{object}{A Seurat object or a list of Seurat objects. This is the primary input that contains single-cell data in the form
of a Seurat object. The function will analyze the specified assay within this object.}

\item{assay}{Character. The name of the assay to be analyzed. Default is "RNA". This assay contains
the expression data to be used for contamination detection.}

\item{group.by}{Character. The metadata column used for grouping cells. Default is "seurat_clusters".
This parameter is used to identify groups within the cells that may have different contamination profiles.}

\item{split.by}{A character string specifying the sample column in the Seurat object's metadata (e.g., "orig.ident"). Default is "orig.ident".}

\item{restriction_factor}{Numeric. A parameter to control the sensitivity of contamination detection.
Values should be between 0 and 1, where a higher value implies a stricter detection. Default is 0.5.}

\item{...}{Additional arguments passed to \code{scCDC::ContaminationDetection} for more customized analysis.}
}
\value{
A list of features or metrics that indicate potential contamination, as returned by
\code{scCDC::ContaminationDetection}.
}
\description{
This function identifies potential contamination features in a given Seurat object. It utilizes the
\code{scCDC::ContaminationDetection} method to detect contamination, making it easier to identify unwanted or erroneous features that might affect
the downstream analysis.
}
\references{
Wang, W., Cen, Y., Lu, Z., Xu, Y., Sun, T., Xiao, Y., Liu, W., Li, J. J., & Wang, C. (2024). scCDC: a computational method for gene-specific contamination detection and correction in single-cell and single-nucleus RNA-seq data. Genome biology, 25(1), 136. https://doi.org/10.1186/s13059-024-03284-wIF: 1
}
