% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{FindSampleMetricsWarning}
\alias{FindSampleMetricsWarning}
\title{Find Metrics Outlier Samples}
\usage{
FindSampleMetricsWarning(
  object,
  sample.by = "orig.ident",
  metrics.by = NULL,
  split.by = NULL,
  type = NULL,
  log = NULL,
  nmads = rep(3, length(metrics.by)),
  aggregate_type = "median",
  color = c("#A6CEE3"),
  return.type = c("table"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object or a data frame.}

\item{sample.by}{A character string specifying the metadata column to group cells by before prediction. Default: "orig.ident".}

\item{metrics.by}{A character vector specifying the metrics to use for outlier detection. Default: c("nCount_RNA", "nFeature_RNA", "nCount_ADT", "nFeature_ADT", "percent.mt", "percent.isotype").}

\item{split.by}{A character string specifying the metadata column to split the data by before prediction. Default: NULL.}

\item{type}{A character vector specifying the type of outlier detection for each metric. Supported values include:
\itemize{
\item "lower": Detect lower outliers.
\item "higher": Detect higher outliers.
}}

\item{log}{A logical vector specifying whether to log-transform the metrics before outlier detection. Default: c(TRUE, TRUE, TRUE, TRUE, FALSE, FALSE).}

\item{nmads}{A numeric vector specifying the number of median absolute deviations (MADs) to use for outlier detection. Default: c(3, 3, 3, 3, 3, 3).}

\item{aggregate_type}{A character string specifying the type of aggregation to use for outlier detection. Supported values include: "median", "mean", "sum". Default:"median"}

\item{color}{A character vector specifying the colors to use for the plot. Default: c('#ed556a', 'white').}

\item{return.type}{A character vector specifying the type of output to return. Supported values include:
\itemize{
\item "table": Return a table of outlier samples.
\item "interactive_table": Return an interactive table of outlier samples.
\item "plot": Return a plot of outlier samples.
}}

\item{verbose}{A logical value. If \code{TRUE} (default), the message will be printed.
If \code{FALSE} , the message will be suppressed (not printed).}
}
\value{
A list containing the outlier samples and an interactive table of outlier samples.
}
\description{
This function identifies outlier samples based on the median absolute deviation (MAD) of the metrics in a Seurat object.
}
