% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMQCData.R
\name{GetSingleCellMQCData}
\alias{GetSingleCellMQCData}
\title{Retrieve Single Cell Multi-omics QC Data from a Seurat Object}
\usage{
GetSingleCellMQCData(object)
}
\arguments{
\item{object}{A Seurat object.}
}
\value{
A list containing the SingleCellMQC data stored within the Seurat object.
}
\description{
Extracts the Single Cell Multi-omics Quality Control (MQC) data stored in the miscellaneous slot of a Seurat object. This function is useful for accessing pre-computed metrics related to quality control of single-cell data.
}
\examples{
\dontrun{
# Assuming `seuratObj` is a Seurat object with MQC data stored
mqcData <- GetSingleCellMQCData(seuratObj)
}
}
\seealso{
\code{\link{AddSingleCellMQCData}} for adding Single Cell MQC data into a Seurat object.
}
