% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untils.R
\name{MergeMatrix}
\alias{MergeMatrix}
\title{Merge Matrices}
\usage{
MergeMatrix(
  matrix_list,
  add_sample_ids = NULL,
  prefix = TRUE,
  cell_id_delimiter = "_",
  ...
)
}
\arguments{
\item{matrix_list}{A list of matrices to merge.}

\item{add_sample_ids}{Optional vector of sample IDs to append to column names.}

\item{prefix}{Logical, whether to add sample IDs as a prefix (default: TRUE).}

\item{cell_id_delimiter}{Character, delimiter to separate sample IDs and column names (default: "_").}

\item{...}{Additional arguments passed to methods.}
}
\value{
A merged matrix of the same class as the input matrices.
}
\description{
This function merges a list of matrices (either \code{dgCMatrix} or \code{IterableMatrix}) into a single matrix.
}
