% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCell.R
\name{PlotADTCutoff}
\alias{PlotADTCutoff}
\title{Plot ADT Doublet Cutoff}
\usage{
PlotADTCutoff(
  object,
  cutoff_table,
  feature1 = NULL,
  feature2 = NULL,
  sample.by = "orig.ident",
  group.by = NULL,
  color = NULL,
  Sample = NULL,
  log.x = T,
  log.y = T,
  ggside = T,
  size = 1
)
}
\arguments{
\item{object}{A Seurat object containing the ADT data.}

\item{cutoff_table}{A data frame containing the cutoff values for the specified features obtained from \code{RunDB_ADT} function.}

\item{feature1}{A character string specifying the first feature to plot.}

\item{feature2}{A character string specifying the second feature to plot.}

\item{sample.by}{A character string specifying the metadata column to use for subsetting by sample. Default is "orig.ident".}

\item{group.by}{A character string specifying the metadata column to use for grouping. Default is NULL.}

\item{color}{A character string specifying the color to use for the plot. Default is NULL.}

\item{Sample}{A character vector specifying the samples to include in the plot. Default is NULL.}

\item{log.x}{A logical value indicating whether to log-transform the x-axis. Default is TRUE.}

\item{log.y}{A logical value indicating whether to log-transform the y-axis. Default is TRUE.}

\item{ggside}{A logical value indicating whether to use ggside for plotting. Default is TRUE.}

\item{size}{A numeric value specifying the size of the points in the plot. Default is 1.}
}
\value{
A ggplot object representing the scatter plot of the specified features with vertical and horizontal lines representing the cutoff values.
}
\description{
This function plots the ADT (Antibody-Derived Tag) cutoff values for specified features in a Seurat object.
}
