% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCell.R
\name{PlotCellMethodFiltration}
\alias{PlotCellMethodFiltration}
\title{Plot cell filtration percentage results from quality control methods}
\usage{
PlotCellMethodFiltration(
  object,
  type.detection = "lq",
  sample.by = "orig.ident",
  color = "#A6CEE3",
  return.type = "plot",
  color.shadow = "white",
  color.text = "black"
)
}
\arguments{
\item{object}{A Seurat object or a data frame. If a Seurat object is provided, the metadata containing quality control information will be extracted for analysis.}

\item{type.detection}{Character string indicating the type of filtration method. Options are \code{"lq"} for low-quality cells or \code{"db"} for doublet cells. Default is \code{"lq"}.}

\item{sample.by}{Character string specifying the column name in the metadata used for grouping cells (e.g., \code{"orig.ident"} for different samples or conditions).}

\item{color}{Character string specifying the color used in the bar plot for representing filtered cells percentage. Default is \code{'#A6CEE3'}.}

\item{return.type}{Character string or vector indicating the type of output to return. Options are \code{"plot"}, \code{"interactive_table"}, or both. Default is \code{"plot"}.}

\item{color.shadow}{Character string specifying the shadow color for the bar plot text. Default is \code{"white"}.}

\item{color.text}{Character string specifying the color of the text in the bar plot. Default is \code{"black"}.}
}
\value{
The function returns either a \code{ggplot} object representing the bar plot or an interactive table, depending on the \code{return.type} specified.
\itemize{
\item If \code{"plot"} is specified, a bar plot showing the percentage of cells filtered out by the selected method for each sample is returned.
\item If \code{"interactive_table"} is specified, an interactive table showing the count and percentage of filtered cells per sample is returned.
\item If both are specified, a list containing both the plot and the interactive table is returned.
}
}
\description{
This function visualizes the percentage of cells filtered out based on quality control methods in a Seurat object.
}
