% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCell.R
\name{PlotCellMethodUpset}
\alias{PlotCellMethodUpset}
\title{Plot overlap of filtered cells using upset plot}
\usage{
PlotCellMethodUpset(
  object,
  type.detection = "lq",
  split.by = NULL,
  nintersects = 20
)
}
\arguments{
\item{object}{A Seurat object or a data frame. If a Seurat object is provided, the metadata containing quality control information will be extracted for analysis.}

\item{type.detection}{Character string indicating the type of filtration method. Options are \code{"lq"} for low-quality cells or \code{"db"} for doublet cells. Default is \code{"lq"}.}

\item{split.by}{Optional. Character string specifying the column in the metadata to split the data by (e.g., \code{"orig.ident"} to visualize separately for each sample). If \code{NULL}, all cells are considered together. Default is \code{NULL}.}

\item{nintersects}{Numeric value indicating the maximum number of intersections to display in the UpSet plot. Default is \code{20}.}
}
\value{
A list of UpSet plots if \code{split.by} is specified; otherwise, a single UpSet plot showing the overlap of filtered cells across different quality control metrics.
}
\description{
This function creates an UpSet plot to visualize the overlap of cells filtered out by different quality control methods (e.g., low quality or doublet cells).
}
