% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCell.R
\name{PlotCellMethodVln}
\alias{PlotCellMethodVln}
\title{Plot quality control metrics with violin plots for filtered cells}
\usage{
PlotCellMethodVln(
  object,
  type.detection = "lq",
  split.by = NULL,
  metrics.by = "nFeature_RNA",
  color = get_colors(2)[2:1]
)
}
\arguments{
\item{object}{A Seurat object or a data frame. If a Seurat object is provided, the metadata containing quality control information will be extracted for visualization.}

\item{type.detection}{Character string indicating the type of filtration method. Options are \code{"lq"} for low-quality cells or \code{"db"} for doublet cells. Default is \code{"lq"}.}

\item{split.by}{Optional. Character string specifying the column in the metadata to split the data by (e.g., \code{"orig.ident"} to visualize separately for each sample). If \code{NULL}, all cells are considered together. Default is \code{NULL}.}

\item{metrics.by}{Character string specifying the quality control metric to plot. Default is \code{"nFeature_RNA"}.}

\item{color}{Vector of colors to use for plotting the different cell types (\code{"Fail"} vs. \code{"Pass"}). Default uses two colors.}
}
\value{
A list of \code{ggplot} violin plot objects if \code{split.by} is specified; otherwise, a single \code{ggplot} object visualizing the chosen quality control metric.
}
\description{
This function generates violin plots to visualize the distribution of selected quality control metrics for cells filtered by different quality control methods (e.g., low quality or doublet cells).
}
