% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_effect.R
\name{PlotCovariateImpact}
\alias{PlotCovariateImpact}
\title{Visualize Covariate Impact on Single-Cell Data}
\usage{
PlotCovariateImpact(
  object,
  plot.type = "cell",
  assay = "RNA",
  variables = NULL,
  pseudobulk.sample.by = "orig.ident",
  return.type = "plot",
  pseudobulk.celltype.by = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data to be analyzed.}

\item{plot.type}{A character string specifying the type of plot to generate. Options include \code{"cell"} and \code{"pseudobulk"} (default: \code{"cell"}).}

\item{assay}{A character string specifying the assay to use for PCA (default: \code{"RNA"}).}

\item{variables}{A character vector specifying the metadata columns to use for variance explained analysis (default: \code{NULL}).}

\item{pseudobulk.sample.by}{A character string specifying the metadata column used for defining samples in the pseudobulk analysis (default: \code{"orig.ident"}).}

\item{return.type}{A character string specifying the output format. Options include \code{"plot"} and \code{"interactive_table"} (default: \code{"plot"}).}

\item{pseudobulk.celltype.by}{A character string specifying the metadata column used to split the data into subsets in the pseudobulk analysis (default: \code{NULL}).}

\item{color}{A vector of colors to use for visualizations (default: \code{NULL}).}

\item{...}{Additional arguments to be passed to the plotting functions.}
}
\value{
A plot or interactive table based on the specified \code{plot.type} and \code{return.type}.
}
\description{
This function evaluates the impact of specified covariates on single-cell data, either at the cell level or pseudobulk level. It supports two main visualization types: covariate variance explained across features (\code{"cell"}) and PCA-based analyses (\code{"pseudobulk"}).
}
\examples{
\dontrun{
# Visualize covariate impact, object is a Seurat object
PlotCovariateImpact(object, plot.type = "cell")
}

}
\seealso{
\code{\link{RunVarExplained}}, \code{\link{PlotVEPerFeature}}, \code{\link{PlotSamplePCA}}
}
