% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFeature.R
\name{PlotFeatureMetrics}
\alias{PlotFeatureMetrics}
\title{Plot Feature Metrics}
\usage{
PlotFeatureMetrics(
  object,
  assay = "RNA",
  type = "pct",
  sample = NULL,
  log.x = T,
  color = "#0072B2",
  ntop = 10,
  ylab = type,
  return.type = "plot"
)
}
\arguments{
\item{object}{A Seurat object processed through the \code{\link{CalculateMetrics}} or \code{\link{CalculateMetricsPerFeature}} function.}

\item{assay}{A character string specifying the assay to analyze. The default is \code{"RNA"}.}

\item{type}{A character string specifying the type of feature metric to visualize.
Valid options include:"pct", "mean", "nCell", "variance", "variance.standardized". Default is "pct".}

\item{sample}{Optional. Character string specifying a sample name to subset the data. If \code{NULL}, the entire dataset is used.}

\item{log.x}{Logical. If \code{TRUE}, the x-axis will be scaled logarithmically. Default is \code{TRUE}.}

\item{color}{Character string specifying the color of the plot (in hex or other color formats). Default is \code{"#0072B2"}.}

\item{ntop}{Numeric. Specifies the number of top features to highlight in the bar plot. Default is \code{10}.}

\item{ylab}{Character string specifying the y-axis label for the top bar plot. Default is the same as the \code{type} parameter.}

\item{return.type}{Character string or vector indicating the type of output to return. Options are \code{"plot"}, \code{"interactive_table"}, or both. Default is \code{"plot"}.}
}
\value{
The function returns either a \code{ggplot} object, an interactive table, or both, depending on the \code{return.type} specified.
\itemize{
\item If \code{"plot"} is specified, a \code{ggplot} object showing the distribution of feature metrics is returned.
\item If \code{"interactive_table"} is specified, an interactive table showing the feature metrics is returned.
\item If both are specified, a list containing both the plot and the interactive table is returned.
}
}
\description{
This function generates plots for visualizing the distribution of specific feature metrics.
}
