% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFeature.R
\name{PlotFeatureMetricsScatter}
\alias{PlotFeatureMetricsScatter}
\title{Plot scatter plot of feature metrics in single cell data}
\usage{
PlotFeatureMetricsScatter(
  object,
  group.by = NULL,
  metrics.by = NULL,
  color = NULL,
  assay = "RNA",
  sample = NULL,
  ggside = T,
  log.x = F,
  log.y = F,
  log.z = F,
  raster.cutoff = 1e+05,
  size = 0.3,
  size.raster = 1,
  ticks = NULL
)
}
\arguments{
\item{object}{A Seurat object processed through the \code{\link{CalculateMetrics}} or \code{\link{CalculateMetricsPerFeature}} function.}

\item{group.by}{Character string specifying a column in the metadata to use for grouping. If provided, points in the plot will be colored by the groups specified in this column.}

\item{metrics.by}{Character vector specifying the metrics to plot. Must contain either 2 or 3 metric names for generating 2D or 3D scatter plots, respectively.}

\item{color}{Optional. A vector of colors used to distinguish different groups.}

\item{assay}{Character string specifying which assay to use. Default is \code{"RNA"}.}

\item{sample}{Optional. Character string specifying a sample to subset the data. If \code{NULL}, all samples are used.}

\item{ggside}{Logical. If \code{TRUE}, includes marginal density plots on the side of the scatter plot. Default is \code{TRUE}.}

\item{log.x}{Logical. If \code{TRUE}, applies a log10 transformation to the x-axis. Default is \code{TRUE}.}

\item{log.y}{Logical. If \code{TRUE}, applies a log10 transformation to the y-axis. Default is \code{TRUE}.}

\item{log.z}{Logical. If \code{TRUE}, applies a log10 transformation to the z-axis in 3D plots. Default is \code{TRUE}.}

\item{raster.cutoff}{Numeric. The number of points at which rasterization is applied to the plot to improve performance. Default is \code{100000}.}

\item{size}{Numeric. Point size for non-rasterized points. Default is \code{0.3}.}

\item{size.raster}{Numeric. Point size for rasterized points. Default is \code{1}.}

\item{ticks}{Numeric vector specifying tick positions for the 3D scatter plot. Optional.}
}
\value{
A ggplot object or list of ggplot objects representing the scatter plot(s).
If multiple assays or samples are provided, a list of plots is returned. For a single assay/sample, a single plot is returned.
}
\description{
This function generates scatter plots to visualize relationships between different feature metrics in single cell data.
It can create either 2D scatter plots or 3D scatter plots based on the number of metrics provided.
}
