% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_other.R
\name{PlotGTEBar}
\alias{PlotGTEBar}
\title{Plot GTE Scores as a Bar Plot}
\usage{
PlotGTEBar(object, color = NULL, ntop = 10)
}
\arguments{
\item{object}{A data frame, typically the output from \code{RunBatchGTEs}. It
should contain at least two columns: 'Feature' (gene names) and a column
representing the GTE scores (usually named after the \code{batch.by} variable
used in \code{RunBatchGTEs}).}

\item{color}{(Optional) A character string specifying the color for the bars.
If \code{NULL}, a default color is used.}

\item{ntop}{An integer specifying the number of top features (genes) to display
in the plot. If \code{NULL} or less than 1, all features will be plotted.
Defaults to 10.}
}
\value{
A \code{ggplot} object representing the bar plot of GTE scores.
}
\description{
This function generates a bar plot to visualize the top genes ranked by
their GTE scores, indicating their association with
a batch effect.
}
\details{
The function sorts the features by their GTE scores in descending order and
selects the top \code{ntop} features. It then creates a horizontal bar plot where
each bar's length corresponds to the GTE score of a gene.
}
