% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{PlotMarkerEnrichKegg}
\alias{PlotMarkerEnrichKegg}
\title{Plot KEGG enrichment by clusterProfiler package.}
\usage{
PlotMarkerEnrichKegg(cosg_marker)
}
\arguments{
\item{cosg_marker}{A data frame containing marker genes, with columns for gene symbols, log fold change, and cluster information.}
}
\value{
A plot representing a dot plot of KEGG pathway enrichment results for each cluster.
}
\description{
This function plots KEGG enrichment results by clusterProfiler package for marker genes identified by COSG.
}
