% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genelist_Benchmark.R
\name{PlotNoiseGeneClusterMetricsVar}
\alias{PlotNoiseGeneClusterMetricsVar}
\title{Plot Variance Explained by Noise Gene Cluster}
\usage{
PlotNoiseGeneClusterMetricsVar(metrics)
}
\arguments{
\item{metrics}{An object containing variance fraction metrics, typically
the \code{Variance_fractions} component from the output of \code{CalNoiseGeneClusterMetrics} function.}
}
\value{
A \code{ggplot} object representing the bar plot of variance explained.
}
\description{
This function visualizes the proportion of variance explained by different
factors. It generates
a stacked bar chart (or dodged bar chart in this specific re-implementation).
}
