% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch.R
\name{PlotReducedDim}
\alias{PlotReducedDim}
\title{Visualize Reduced Dimensional Data}
\usage{
PlotReducedDim(
  object,
  group.by = "seurat_clusters",
  process.pseudobulk = "pca",
  reduction = "rna.umap",
  split.by = NULL,
  ncol = NULL,
  ggside = F,
  color = NULL,
  guide.nrow = 10,
  raster.cutoff = 1e+05,
  pca.label = NULL,
  size = 0.3
)
}
\arguments{
\item{object}{A Seurat object directly, or a list/object (typically output
from \code{RunPseudobulkData})
for pseudobulk-level PCA plotting.}

\item{group.by}{A character string specifying the metadata column used
for coloring points (e.g., "seurat_clusters").}

\item{process.pseudobulk}{A character string. If "pca" and \code{object} is
a pseudobulk data list, it will first run PCA on the pseudobulk matrix
and then plot its PCA embeddings. If \code{object} is a Seurat object, this
parameter is ignored. Defaults to "pca".}

\item{reduction}{A character string specifying the name of the dimensionality
reduction to plot (e.g., "umap", "pca"). For single-cell Seurat object,
this typically refers to existing reductions. For pseudobulk data,
if \code{process.pseudobulk="pca"}, this will automatically be "pca".
Defaults to "rna.umap".}

\item{split.by}{(Optional) A character string specifying a metadata column
by which to facet the plots. Defaults to \code{NULL}.}

\item{ncol}{(Optional) An integer specifying the number of columns for the
facet plot. If \code{NULL}, it will be automatically determined.}

\item{ggside}{A logical value. If \code{TRUE}, it attempts to add marginal plots
using \code{ggside} (if installed and loaded). Defaults to \code{FALSE}.}

\item{color}{(Optional) A character vector of colors to use for \code{group.by}.
If \code{NULL}, default colors are assigned.}

\item{guide.nrow}{An integer specifying the number of rows for the legend guide.
Defaults to 10.}

\item{raster.cutoff}{An integer specifying the number of points above which
data will be rasterized to improve plot rendering performance. Points below
this threshold will be plotted as individual points. Defaults to 100000.}

\item{pca.label}{A character string specifying the metadata column used for label.}

\item{size}{A numeric value specifying the size of the plotted points.
Defaults to 0.3.}
}
\value{
A \code{ggplot} object, or a named list of \code{ggplot} objects if \code{object$pseudobulk_mat}
is a list (e.g., pseudobulk PCA plots per cluster).
}
\description{
This function generates scatter plots of reduced dimensional data, either
at the cell level (from a single-cell object) or at the pseudobulk PCA level.
}
