% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample.R
\name{PlotSampleCellTypePCT}
\alias{PlotSampleCellTypePCT}
\title{Plot Sample Cell Type Percentage}
\usage{
PlotSampleCellTypePCT(
  object,
  sample.by = "orig.ident",
  celltype.by = "ScType",
  color.stackbar = NULL,
  color.bar = "#A6CEE3",
  plot.type = "stackbar",
  nrow.bar = NULL,
  ncol.bar = NULL,
  levels = NULL,
  color.text.bar = "black",
  color.shadow.bar = "white",
  return.type = "plot",
  maxWidth = 85
)
}
\arguments{
\item{object}{A Seurat object or a data frame containing the metadata information.}

\item{sample.by}{The column name in \code{object} that contains the sample information.}

\item{celltype.by}{The column name in \code{object} that contains the cell type information.}

\item{color.stackbar}{A vector of colors to use for the stacked bar plot. The length of the vector should be equal to the number of unique cell types. Default is NULL.}

\item{color.bar}{The color to use for the bar plot. Default is "#A6CEE3".}

\item{plot.type}{The type of plot to generate. Supported values include "stackbar" and "bar". Default is "stackbar".}

\item{nrow.bar}{The number of rows for the bar plot, if \code{plot.type} is "bar". Default is NULL.}

\item{ncol.bar}{The number of columns for the bar plot, if \code{plot.type} is "bar". Default is NULL.}

\item{levels}{The order of the samples to be displayed on the x-axis. If NULL, the order is determined by the order of the unique values in the \code{sample.by} column.}

\item{color.text.bar}{The color of the text annotations on the bar plot. Default is "black"}

\item{color.shadow.bar}{The color of the shadow behind the text annotations on the bar plot. Default is "white"}

\item{return.type}{The type of output to return. Supported values include "plot" and "interactive_table".}

\item{maxWidth}{The maximum width of the interactive table. Default is 85.}
}
\value{
Depending on the input parameters, this function can return a ggplot2::ggplot object for direct visualization in R, or an interactive data table that can be used for custom analyses or exported as CSV.
}
\description{
This function generates a bar plot or stacked bar plot showing the percentage of cell types in each sample.
}
\examples{
\dontrun{
# Assuming 'seurat_obj' is a Seurat object with metadata information:
PlotSampleCellTypePCT(seurat_obj, sample.by = "orig.ident", celltype.by = "ScType")
}
}
\seealso{
\link{ShowSampleMetricsName}
}
