% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample.R
\name{PlotSampleLabel}
\alias{PlotSampleLabel}
\title{Plot Sample Label}
\usage{
PlotSampleLabel(
  object,
  sample.by = "orig.ident",
  assay = "RNA",
  feature_list = list(Female = "XIST", Male = c("DDX3Y", "UTY", "RPS4Y1")),
  color = c("#A6CEE3", "#B2DF8A"),
  return.type = "plot"
)
}
\arguments{
\item{object}{A Seurat object or a data frame containing the percentage of features in each sample.}

\item{sample.by}{The column name in the meta.data slot of the Seurat object to split the data by.}

\item{assay}{The assay to use for the feature calculation.}

\item{feature_list}{A list of features to calculate the percentage of. The list should be named, with each element containing a vector of feature names.}

\item{color}{A vector of colors to use for the bar fill, corresponding to the unique values in \code{color.by}.}

\item{return.type}{A character vector indicating the type of output to return. Supported values include "plot" and "interactive_table".}
}
\value{
A ggplot2::ggplot object representing the bar plot.
}
\description{
The function is used to plot sample labels based on the percentage of features in each sample.
}
