% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample.R
\name{PlotSampleMetrics}
\alias{PlotSampleMetrics}
\title{Plot Sample Metrics}
\usage{
PlotSampleMetrics(
  object,
  metrics = c("nCell", "nGene_RNA", "nPro_ADT", "nChain_TCR", "nCell_TCR", "TCR\%",
    "nChain_BCR", "nCell_BCR", "BCR\%"),
  type = "count",
  metrics_rename = NULL,
  color = "#A6CEE3",
  color.shadow = "white",
  reorder = NULL,
  color.text = "black",
  return.type = "plot",
  csv.name = "count",
  elementId = csv.name,
  table.subtitle = "Metrics",
  maxWidth = 85
)
}
\arguments{
\item{object}{A Seurat object.}

\item{metrics}{A character vector specifying the metrics to plot.}

\item{type}{The type of metrics to show. Supported values include "count", "summary", and "Metrics_10x". Default is "count".}

\item{metrics_rename}{A character vector specifying the names to use for the metrics in the plot. Default is NULL.}

\item{color}{A character vector specifying the colors to use for the bar plot. Default is "#A6CEE3".}

\item{color.shadow}{The color of the shadow behind the text annotations on the bar plot. Default is "white".}

\item{reorder}{A character vector specifying the order of the samples to be displayed on the x-axis. Default is NULL.}

\item{color.text}{The color of the text annotations on the bar plot. Default is "black".}

\item{return.type}{The type of output to return. Supported values include "plot" and "interactive_table". Default is "plot".}

\item{csv.name}{The name of the CSV file to download. Default is "count".}

\item{elementId}{The ID of the element to use for the interactive table.}

\item{table.subtitle}{The subtitle to add to the interactive table. Default is "Metrics".}

\item{maxWidth}{The maximum width of the interactive table. Default is 85.}
}
\value{
Depending on the input parameters, this function can return a ggplot2::ggplot object for direct visualization in R, or an interactive data table that can be used for custom analyses or exported as CSV.
}
\description{
This function generates a bar plot or interactive data table showing the specified metrics for each sample in the Seurat object.
}
\examples{
\dontrun{
# Assuming 'seurat_obj' is a Seurat object:
# Visualisation of the count number of metrics
PlotSampleMetrics(seurat_obj, return.type = "plot")
PlotSampleMetrics(pbmc, type="count", metrics=c("nCell", "nGene_RNA"), return.type = "plot" )

# Visualisation of the summary of percent.mt
PlotSampleMetrics(seurat_obj, type="summary", metrics=c("percent.mt"), return.type = "plot"  )

# Visualisation  of the Metrics_10x
PlotSampleMetrics(seurat_obj, type="Metrics_10x",  return.type = "interactive_table"  )

# return an interactive table
PlotSampleMetrics(seurat_obj, return.type = "interactive_table")

}
}
\seealso{
\link{ShowSampleMetricsName}
}
