% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_effect.R
\name{PlotSamplePCA}
\alias{PlotSamplePCA}
\title{Visualize PCA Results on Pseudobulk Data}
\usage{
PlotSamplePCA(
  object,
  assay = "RNA",
  sample.by = "orig.ident",
  svd.variable.by = NULL,
  pca.group.by = "orig.ident",
  celltype.by = NULL,
  do.aggregate = T,
  ntop = 2000,
  color = NULL,
  pca.size.point = 3,
  pca.size.text = 3,
  pca.Legend.lab = NULL,
  svd.maxPCs = 3,
  plot.type = "pca",
  return.type = "plot"
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data to be analyzed.}

\item{assay}{A character string specifying the assay to use for PCA (default: \code{"RNA"}).}

\item{sample.by}{A character string specifying the metadata column used for defining samples (default: \code{"orig.ident"}).}

\item{svd.variable.by}{A character vector of metadata columns to use for SVD analysis (default: \code{NULL}).}

\item{pca.group.by}{A character string specifying the metadata column used for grouping cells in the PCA plot (default: \code{"orig.ident"}).}

\item{celltype.by}{A character string specifying the metadata column used to split the data into subsets (default: \code{NULL}).}

\item{do.aggregate}{Logical, whether to aggregate counts across groups before PCA (default: \code{TRUE}).}

\item{ntop}{An integer specifying the number of top variable features to use for PCA (default: \code{2000}).}

\item{color}{A vector of colors to use for visualizations (default: \code{NULL}).}

\item{pca.size.point}{A numeric value specifying the point size for PCA plots (default: \code{3}).}

\item{pca.size.text}{A numeric value specifying the text size for PCA plots (default: \code{3}).}

\item{pca.Legend.lab}{A character string specifying the legend label for PCA plots (default: \code{NULL}).}

\item{svd.maxPCs}{An integer specifying the maximum number of PCs to include in the SVD plot (default: \code{3}).}

\item{plot.type}{A character vector specifying the types of plots to generate. Options include \code{"pca"}, \code{"svd"}, and \code{"explanatory"} (default: \code{"pca"}).}

\item{return.type}{A character vector specifying the output format. Options include \code{"plot"}, \code{"interactive_table"}, and \code{"table"} (default: \code{"plot"}).}
}
\value{
A list containing plots, tables, or interactive tables based on the specified \code{plot.type} and \code{return.type}.
}
\description{
This function generates PCA plots, SVD significance analysis, or explanatory variance plots using pseudobulk data. It allows for splitting data by a specified variable, aggregating counts, and visualizing the results.
}
\examples{
\dontrun{
# Generate PCA plots
object <- CreateSeuratObject(counts = matrix(rpois(1000, lambda = 10), nrow = 100, ncol = 10))
pca_results <- PlotSamplePCA(object, sample.by = "orig.ident", plot.type = c("pca", "svd"))
}

}
