% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample.R
\name{PlotSampleVDJ}
\alias{PlotSampleVDJ}
\title{Plot V(D)J Sample Statistics}
\usage{
PlotSampleVDJ(
  object,
  color = NULL,
  type = "pct",
  sample_name = names(object),
  size = 4,
  return.type = "plot",
  return.scRepertoire = F,
  scRepertoire_data = NULL
)
}
\arguments{
\item{object}{A Seurat object or a list containing the original VDJ table for each sample.
\itemize{
\item If type is one of "pct" or "subtype", a Seurat object processed with \code{\link{CalculateMetricsPerSample}} or \code{\link{CalculateMetrics}} is required as input.
\item If type is one of "CDR3", "clonalQuant", "clonalOverlap", or "geneUsage", lists obtained by the \code{\link{Read10XData}} function are required as input.
}
The input ideally contains V(D)J recombination data processed through the appropriate \code{scRepertoire} or similar analysis pipelines.}

\item{color}{A character vector specifying colors to be used in the plot. The length of the vector should match the sample size.}

\item{type}{A character string indicating the type of V(D)J visualization or data to return. Supported types include:
\itemize{
\item "pct": for the percentage of chain type, Seurat object as input.
\item "subtype": for the percentage of chain subtype, Seurat object as input.
\item "CDR3": for the length distribution of the CDR3 sequences, raw list of "filtered_contig_annotations" data as input, details see \code{\link[scRepertoire]{clonalLength}} function from \code{scRepertoire} package.
\item "clonalQuant": for quantifying unique clonotypes, raw list of "filtered_contig_annotations" data as input, details see \code{\link[scRepertoire]{clonalQuant}} function from \code{scRepertoire} package.
\item "clonalOverlap": for measures of similarity between samples, raw list of "filtered_contig_annotations" data as input, details see \code{\link[scRepertoire]{clonalOverlap}} function from \code{scRepertoire} package.
\item "geneUsage": for the proportion of V gene usage, raw list of "filtered_contig_annotations" data as input, details see \code{\link[scRepertoire]{percentGenes}} function from \code{scRepertoire} package.
}}

\item{sample_name}{Optional; a character vector specifying sample names, applicable only for the \code{type} parameter values: "CDR3", "clonalQuant", "clonalOverlap", or "geneUsage".}

\item{size}{Numeric; the size of points in scatter plots. Defaults to 4.}

\item{return.type}{Character string specifying the output type, either "plot" or "interactive_table". Defaults to "plot".}

\item{return.scRepertoire}{A logical indicating whether to return the list obtained from \code{scRepertoire::combineTCR} or \code{scRepertoire::combineBCR}, applicable only for the \code{type} parameter values: "CDR3", "clonalQuant", "clonalOverlap", or "geneUsage".}

\item{scRepertoire_data}{A list processed by \code{scRepertoire::combineTCR} or \code{scRepertoire::combineBCR}, applicable only for the \code{type} parameter values: "CDR3", "clonalQuant", "clonalOverlap", or "geneUsage". Default is NULL.}
}
\value{
Depending on the input parameters, this function returns either a ggplot2::ggplot object for direct visualization in R or an interactive data table that can be used for custom analyses or exported as a CSV. The specific output is determined by the \code{type} and \code{return.interactive_table} parameters.
}
\description{
Generates a plot or interactive data table for V(D)J metrics, subtype, CDR3 length, clonal quantification, clonal overlap, and gene usage.
This function is part of a suite of tools designed for exploring immunological data within single-cell datasets, focusing on V(D)J recombination data.
}
