% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariancePartition.R
\name{PlotVarPartVln}
\alias{PlotVarPartVln}
\title{Plot Variance Partition Results as Violin Plots}
\usage{
PlotVarPartVln(object, color = NULL, do.split = F)
}
\arguments{
\item{object}{A data frame, typically the output of \code{RunVarPartMetadata},
\code{RunVarPartPseudobulk}, or \code{RunVarPartPseudobulkPCA}. It should contain
at least 'Feature' and 'Group' columns (if applicable), and columns
representing the variance components (e.g., 'donor_id', 'batch').}

\item{color}{(Optional) A character vector of colors to be used for filling
the violin plots. If \code{NULL}, default colors are assigned.}

\item{do.split}{A logical value. If \code{TRUE}, it expects the \code{object} to
already contain a 'Group' column and will facet the plots by this 'Group'
column. If \code{FALSE} and multiple groups are present, it will generate a
separate plot for each group.}
}
\value{
A \code{ggplot} object (or a list of \code{ggplot} objects if \code{do.split} is \code{FALSE}
and multiple groups are present) representing the violin plots of variance explained.
}
\description{
Generates violin plots to visualize the proportion of variance explained by
different factors across features (genes, PCs, or metadata features).
}
