% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_data.R
\name{Read10XMetrics}
\alias{Read10XMetrics}
\title{Read 10X QC Metrics}
\usage{
Read10XMetrics(file_path, sample_name = NULL)
}
\arguments{
\item{file_path}{A named vector or list of file paths to the QC metrics files.}

\item{sample_name}{An optional vector of sample names corresponding to the files in \code{file_path}.
If \code{file_path} is unnamed and \code{sample_name} is not provided, sample names will be generated automatically as "sample_1", "sample_2", etc.}
}
\value{
A data frame combining the processed QC metrics for all samples.
}
\description{
This function reads 10X Genomics QC metrics files and processes them based on the data type.
It supports multiple file types, including multi-library metrics, VDJ metrics, and count metrics.
}
