% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_data.R
\name{ReadVDJ}
\alias{ReadVDJ}
\title{Read V(D)J Sequencing Data From 10X.}
\usage{
ReadVDJ(dir10x, sample = seq_along(dir10x))
}
\arguments{
\item{dir10x}{Directory containing filtered_contig_annotations.csv. e.g. "/data/SingleCellMQC/CellRanger/TP1/vdj_t/filtered_contig_annotations.csv"}

\item{sample}{Optional; sample identifiers. If not specified, defaults to sequence along dir10x.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A list of contig tables, each list associated with a sample.
}\if{html}{\out{</div>}}
}
\description{
Reads V(D)J sequencing data for TCR or BCR provided by 10X and combines them into a list.
}
