% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch.R
\name{RunBatchGTE}
\alias{RunBatchGTE}
\title{Quantify Feature-level Technical Effects (GTE) for Single-cell Data}
\usage{
RunBatchGTE(object, assay = "RNA", slot = "data", group.by, batch.by)
}
\arguments{
\item{object}{A single-cell object.}

\item{assay}{A character string specifying the assay to use (e.g., "RNA")
if \code{object} is a single-cell object. Defaults to "RNA".}

\item{slot}{A character string specifying the slot to use for expression
data (e.g., "data", "counts") if \code{object} is a single-cell object.
Defaults to "data".}

\item{group.by}{(Optional) A character string specifying the metadata column
representing biological groups (e.g., "cell_type", "condition") to be
controlled for in the GTE calculation. This helps isolate batch effects
from true biological variation.}

\item{batch.by}{A character string specifying the metadata column representing
the technical factor (e.g., "batch_id", "sequencing_run") whose influence
is to be quantified.}
}
\value{
A \code{data.frame}
\itemize{
\item \code{Feature}: Gene or protein names.
\item \code{GTE_Overall}: The overall GTE score for each gene, quantifying
the total influence of the \code{batch.by} variable.
\item Additional columns: For each subgroup defined by \code{group.by}, if
provided, a column (e.g., "CellTypeA", "CellTypeB") will represent
the GTE score specific to that subgroup.
}
}
\description{
This function calculates feature-level Technical Effects (GTE) scores for
individual genes (or other features) in single-cell expression data,
measuring the influence of user-defined technical factors (batch variables).
It is an adapted and extended implementation of the methodology originally
presented in the \code{GTEs::Run.GroupTechEffects} function, designed to offer
broader compatibility with various single-cell data formats, such as
\code{BPCells} matrices, \code{Seurat} object.
The core GTE principle is based on the work by Zhou et al. (2025).
}
\references{
Zhou Y, Sheng Q, Wang G, Xu L, Jin S. Quantifying batch effects
for individual genes in single-cell data. Nat Comput Sci. 2025 Aug;5(8):612-660.
doi: 10.1038/s43588-025-00824-7. Epub 2025 Jun 27. PMID: 40579473.
}
\seealso{
\code{\link[GTEs]{Run.GroupTechEffects}}
}
