% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch.R
\name{RunBatchTest}
\alias{RunBatchTest}
\title{Run Batch Test}
\usage{
RunBatchTest(
  object,
  sample.by = "orig.ident",
  cluster.by = "seurat_clusters",
  k = 3,
  seed = 1,
  n = 100
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data to be analyzed.}

\item{sample.by}{A character string specifying the metadata column used for defining samples (default: \code{"orig.ident"}).}

\item{cluster.by}{A character string specifying the metadata column used for grouping cells (default: \code{"seurat_clusters"}).}

\item{k}{An integer specifying the number of clusters to use for batch testing (default: \code{3}).}

\item{seed}{An integer specifying the seed for random sampling (default: \code{1}).}

\item{n}{An integer specifying the number of samples to use for batch testing (default: \code{100}).}
}
\value{
A list containing the correlation matrix, Kendall W results, and the qbinom results.
}
\description{
This function performs a batch test on a Seurat object to evaluate the presence of batch effects. It calculates the Kendall W statistic, pairwise Kendall correlations, and the number of negative correlations between samples.
The function also checks for severe batch effects based on the results.
}
