% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_QC.R
\name{RunCorrection}
\alias{RunCorrection}
\title{A wrapper function for contamination correction}
\usage{
RunCorrection(
  object,
  method = NULL,
  group.by = "seurat_clusters",
  features = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object. The input data must be a valid Seurat object to proceed with correction.}

\item{method}{Character string specifying the correction method to use. Options include:
\itemize{
\item \code{"DecontX"}: Runs the \code{RunCorrection_DecontX()} function.
\item \code{"DecontPro"}: Runs the \code{RunCorrection_DecontPro()} function.
\item \code{"scCDC"}: Runs the \code{RunCorrection_scCDC()} function.
}
If not specified, the function will return an error indicating that a valid method must be chosen.}

\item{group.by}{Character string specifying the grouping variable (default is \code{"seurat_clusters"}). This parameter is mainly used when the method is \code{"DecontPro"}, allowing you to perform the correction in a grouped manner.}

\item{features}{Optional vector of feature names. Used specifically by the \code{scCDC} method to indicate which features to consider during correction.}

\item{...}{Additional arguments passed to the respective correction function. These arguments will depend on the selected correction method.}
}
\value{
A Seurat object with corrected data, depending on the chosen method.
}
\description{
This function performs contamination correction on a Seurat object using the specified method. The available methods include \code{DecontX}, \code{DecontPro}, and \code{scCDC}.
}
