% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_QC.R
\name{RunCorrection_DecontPro}
\alias{RunCorrection_DecontPro}
\title{Perform the contamination correction by DecontPro}
\usage{
RunCorrection_DecontPro(object, group.by = "seurat_clusters", ...)
}
\arguments{
\item{object}{A Seurat object.}

\item{group.by}{The column name in metadata that contains the clusters information.}

\item{...}{Additional arguments passed to \code{decontX::decontPro} for more customized analysis.}
}
\value{
A Seurat object contains additional corrected assay.
}
\description{
Perform the contamination correction by DecontPro
}
