% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_QC.R
\name{RunCorrection_scCDC}
\alias{RunCorrection_scCDC}
\title{Perform Contamination Correction Using scCDC}
\usage{
RunCorrection_scCDC(
  object,
  assay = "RNA",
  features = NULL,
  split.by = "orig.ident",
  do.merge = T,
  group.by = "seurat_clusters",
  ...
)
}
\arguments{
\item{object}{A Seurat object or a list of Seurat objects.}

\item{assay}{Character. The name of the assay to be analyzed. Default is "RNA".}

\item{features}{Vector or a list of features (genes) to correct. If NULL, all features are used. Default is NULL.}

\item{split.by}{Character. The column name in the metadata used to split the object into multiple samples (e.g., "orig.ident"). Default is "orig.ident"}

\item{do.merge}{Logical. Whether to merge the corrected objects into a single Seurat object. Default is TRUE.}

\item{group.by}{A character string specifying the metadata column used for grouping cells (default: \code{"seurat_clusters"}).}

\item{...}{Additional arguments passed to \code{scCDC::ContaminationCorrection} for customized analysis.}
}
\value{
A Seurat object containing the corrected data in a new assay named "Corrected". If \code{do.merge} is TRUE and multiple samples are provided, a merged Seurat object is returned.
}
\description{
This function applies the scCDC contamination correction method to a Seurat object or a list of Seurat objects.
It supports multi-sample analysis and returns a Seurat object with a new "Corrected" assay.
}
