% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunDbt}
\alias{RunDbt}
\title{Doublet Detection in Single-Cell Data}
\usage{
RunDbt(
  object,
  methods = c("hybrid", "bcds", "cxds", "scDblFinder"),
  add.Seurat = T,
  split.by = "orig.ident",
  db.rate = NULL,
  split.db.rate.1000 = 0.008,
  feature.RNA1 = NULL,
  feature.RNA2 = NULL,
  feature.ADT1 = NULL,
  feature.ADT2 = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{methods}{A character vector specifying which doublet detection methods to apply. Supported methods include: "hybrid", "bcds", "cxds", "scDblFinder", "DoubletFinder", "VDJ", and "ADT".}

\item{add.Seurat}{Logical; if TRUE, adds the doublet detection results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the doublet detection results.}

\item{split.by}{A character string specifying the column in the Seurat object's metadata to use for stratifying the dataset before applying doublet detection (e.g., "orig.ident").}

\item{db.rate}{The expected doublet rate, if known; otherwise, it will be estimated 0.08*nCell/10000. Not applicable to VDJ and ADT methods.}

\item{split.db.rate.1000}{The expected doublet rate per 1000 cells. Default is 0.008, (so 0.016 among 1600 cells).}

\item{feature.RNA1}{Character vector specifying RNA features for ADT doublet detection method.}

\item{feature.RNA2}{Character vector specifying additional RNA features for ADT doublet detection method.}

\item{feature.ADT1}{Character vector specifying ADT features for ADT doublet detection method.}

\item{feature.ADT2}{Character vector specifying additional ADT features for ADT doublet detection method.}

\item{...}{Additional arguments passed to individual doublet detection methods. \code{\link{RunDbt_hybrid}}; \code{\link{RunDbt_bcds}}; \code{\link{RunDbt_cxds}}; \code{\link{RunDbt_scDblFinder}}; \code{\link{RunDbt_DoubletFinder}}; \code{\link{RunDbt_VDJ}}; \code{\link{RunDbt_ADT}};}
}
\value{
If \code{add.Seurat} is TRUE, returns the modified Seurat object with doublet detection results added. Otherwise, returns a list or dataframe with the doublet scores and classifications.
}
\description{
A wrapper function that applies various doublet detection methods to a single-cell dataset and optionally adds results to Seurat object.
}
\seealso{
\code{\link{RunDbt_hybrid}}; \code{\link{RunDbt_bcds}}; \code{\link{RunDbt_cxds}}; \code{\link{RunDbt_scDblFinder}}; \code{\link{RunDbt_DoubletFinder}}; \code{\link{RunDbt_VDJ}}; \code{\link{RunDbt_ADT}};
}
