% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{RunDbt_ADT}
\alias{RunDbt_ADT}
\title{Doublet Detection (ADT) with mutually exclusive CITE-seq markers in CITE-seq data.}
\usage{
RunDbt_ADT(
  object,
  split.by = "orig.ident",
  feature1 = NULL,
  feature2 = NULL,
  method.predict = c("NB", "ROC", "Logit", "LDA"),
  resolution = 1.5,
  dims = 1:20,
  do.correct = F,
  preprocess = "adt.umap",
  return.cutoff = F,
  add.Seurat = T,
  BPtmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing assays: ADT or RNA+ADT.}

\item{split.by}{A character string specifying the column in the Seurat object's metadata to use for stratifying the dataset before applying doublet detection (e.g., "orig.ident").}

\item{feature1}{A character vector of the names of the mutually exclusive CITE-seq markers to use for doublet detection.}

\item{feature2}{A character vector of the names of the mutually exclusive CITE-seq markers to use for doublet detection.}

\item{method.predict}{Character; the method used to predict doublets. Default is "NB". Options are "NB", "ROC", "Logit", and "LDA".}

\item{resolution}{Numeric; the resolution parameter used for clustering. Default is 1.5.}

\item{dims}{Numeric; the dimensions to use for clustering. Default is 1:20.}

\item{do.correct}{Logical; if TRUE, performs correction for batch effects before doublet detection. Default is FALSE.}

\item{preprocess}{Character; name of the preprocessing method to use. Default is "adt.umap".}

\item{return.cutoff}{Logical; if TRUE, returns the cutoff values used for doublet detection. Default is FALSE.}

\item{add.Seurat}{Logical; if TRUE, adds the doublet detection results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the doublet detection results.}

\item{BPtmpdir}{Temporary directory for BPCells matrix processing. Default is "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/".}

\item{...}{Additional arguments.}
}
\value{
If add.Seurat is TRUE, returns the modified Seurat object with doublet detection results added. Otherwise, returns a data frame with the doublet classifications.
}
\description{
This function based on mutually exclusive CITE-seq markers for doublet detection in CITE-seq data.
}
