% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{RunDbt_bcds}
\alias{RunDbt_bcds}
\title{Doublet Detection with bcds}
\usage{
RunDbt_bcds(
  object,
  db.rate = NULL,
  split.db.rate.1000 = 0.008,
  split.by = "orig.ident",
  add.Seurat = T,
  BPtmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  ...
)
}
\arguments{
\item{object}{A single-cell object (e.g., Seurat object).}

\item{db.rate}{The expected doublet rate, if known; otherwise, it will be estimated 0.08*nCell/10000.}

\item{split.db.rate.1000}{The expected doublet rate per 1000 cells. Default is 0.008.}

\item{split.by}{Column name by which to group the data before calculating statistics.}

\item{add.Seurat}{Logical; if TRUE, adds the bcds results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the bcds results.}

\item{BPtmpdir}{Temporary directory for BPCells matrix processing. Default is "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/".}

\item{...}{Additional arguments to be passed to \code{\link[scds]{bcds}} function from the \code{scds} package.}
}
\value{
If \code{add.Seurat} is TRUE, a Seurat object with appended metadata; otherwise, a data frame with the bcds results.
}
\description{
Performs doublet detection using a bcds method in the \code{scds} package on a Seurat object.
}
\details{
The results contained bcds method scores for each cell and classification as 'Pass' or 'Fail'.
}
\references{
Bais, Abha S, and Dennis Kostka. “scds: computational annotation of doublets in single-cell RNA sequencing data.” Bioinformatics (Oxford, England) vol. 36,4 (2020): 1150-1158. doi:10.1093/bioinformatics/btz698
}
\seealso{
\code{\link{RunDbt}}
}
