% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunDbt_DoubletFinder}
\alias{RunDbt_DoubletFinder}
\title{Doublet Detection with DoubletFinder}
\usage{
RunDbt_DoubletFinder(
  object,
  split.by = "orig.ident",
  add.Seurat = TRUE,
  db.rate = NULL,
  split.db.rate.1000 = 0.008,
  PCs = 10,
  BPtmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  ...
)
}
\arguments{
\item{object}{A Seurat object. The input must be a valid Seurat object containing single-cell RNA-seq data.}

\item{split.by}{Character. The name of the metadata column used to split the Seurat object. Default is "orig.ident".}

\item{add.Seurat}{Logical. If TRUE, the function will add the doublet information back to the input Seurat object
as new metadata fields. If FALSE, the function will return a data frame containing the doublet detection results. Default is TRUE.}

\item{db.rate}{Numeric. The doublet rate score used to estimate the number of expected doublets.
Default is calculated as 0.08 * ncol(object)/10000. If NULL, this default formula will be used.}

\item{split.db.rate.1000}{Numeric. The expected doublet rate per 1000 cells. Default is 0.008, (so 0.016 among 1600 cells).}

\item{PCs}{Numeric. The number of principal components. Default is 10.}

\item{BPtmpdir}{Temporary directory for BPCells matrix processing. Default is "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/".}

\item{...}{Additional arguments.}
}
\value{
If add.Seurat is TRUE, returns the input Seurat object with additional metadata fields for doublet detection.
Otherwise, returns a data frame with doublet scores and classifications for each cell.
}
\description{
This function uses the DoubletFinder package to detect doublets within a Seurat object.
}
