% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{RunDbt_scDblFinder}
\alias{RunDbt_scDblFinder}
\title{Doublet Detection with scDblFinder}
\usage{
RunDbt_scDblFinder(
  object,
  do.topscore = T,
  db.rate = NULL,
  split.db.rate.1000 = 0.008,
  split.by = "orig.ident",
  seed = 1,
  add.Seurat = T,
  BPtmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  ...
)
}
\arguments{
\item{object}{A single-cell object (e.g., Seurat object).}

\item{do.topscore}{Logical; if TRUE, uses a top scoring method to identify doublets based on scDblFinder scores.}

\item{db.rate}{Numeric; the expected doublet rate for the top score method. Default is calculated as 0.08 * number of cells / 10000.}

\item{split.db.rate.1000}{The expected doublet rate per 1000 cells. Default is 0.008.}

\item{split.by}{A character string specifying the column in the Seurat object's metadata to use for stratifying the dataset before applying doublet detection (e.g., "orig.ident"). Default is "orig.ident".}

\item{seed}{Random seed for reproducibility. Default is 1.}

\item{add.Seurat}{Logical; if TRUE, adds the doublet detection results as metadata to the Seurat object and stores it in the \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the doublet detection results. Default is FALSE.}

\item{BPtmpdir}{Temporary directory for BPCells matrix processing. Default is "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/".}

\item{...}{Additional arguments to be passed to \code{\link[scDblFinder]{scDblFinder}} function from the \code{scDblFinder} package.}
}
\value{
If \code{add.Seurat} is TRUE, returns the modified Seurat object with doublet detection results added. Otherwise, returns a list or data.frame with the doublet scores and classifications.
}
\description{
Performs doublet detection using a scDblFinder method in the \code{scDblFinder} package on a Seurat object.
}
\details{
The results contained scDblFinder method scores for each cell and classification as 'Pass' or 'Fail'. 'Fail' represents doublets.
}
\references{
Germain, Pierre-Luc et al. “Doublet identification in single-cell sequencing data using scDblFinder.” F1000Research vol. 10 979. 28 Sep. 2021, doi:10.12688/f1000research.73600.2
}
\seealso{
\code{\link{RunDbt}}
}
