% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublet.R
\name{RunDbt_VDJ}
\alias{RunDbt_VDJ}
\title{VDJ Doublet Detection in single cell TCR and/or BCR data.}
\usage{
RunDbt_VDJ(object, add.Seurat = T, ...)
}
\arguments{
\item{object}{A Seurat object.}

\item{add.Seurat}{Logical; if TRUE, adds the doublet detection results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the doublet detection results.}

\item{...}{Additional arguments}
}
\value{
If \code{add.Seurat} is TRUE, returns the modified Seurat object with doublet detection results added. Otherwise, returns a list or data.frame with the doublet classifications.
}
\description{
Performs doublet detection based on VDJ sequencing information.
}
\details{
The classification as 'Pass' or 'Fail'. 'Fail' represents doublets.
'multichain' and 'ambiguous' in the \code{chain_pair} column of the \code{\link{CalculateMetricsPerCell}} function are identified as doublets.
}
\examples{
\dontrun{
# Assuming 'seuratObj' is a Seurat object
results_VDJ <- RunDbt_VDJ(object = seurat_obj, add.Seurat = TRUE)
}
}
\seealso{
\code{\link{RunDbt}}
}
