% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowQuality.R
\name{RunLQ}
\alias{RunLQ}
\title{Low-Quality Cells Detection}
\usage{
RunLQ(
  object,
  methods = c("ddqc", "miQC", "MAD"),
  add.Seurat = T,
  split.by = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{methods}{A character vector specifying the quality assessment methods to apply (one or more). Available method options are:
\itemize{
\item "MAD": Outliers based on the median absolute deviation. Suitable for RNA and ADT.
\item "ddqc": Only RNA. An unsupervised adaptive QC framework to perform flexible and data-driven QC at the level of cell types while retaining critical biological insights and improved power for downstream analysis. (Subramanian, Ayshwarya et al., Genome biology, 2022, PMID: 36575523)
\item "miQC": Only RNA. An adaptive probabilistic framework for quality control of single-cell RNA-sequencing data. (Hippen, Ariel A et al., PLoS computational biology, 2021, PMID: 34428202)
\item "fixed": Fixed cutoffs for RNA and ADT metrics.
}
The default is all methods.}

\item{add.Seurat}{Logical; if TRUE, adds the low-quality cell detection results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the low-quality cell detection results.}

\item{split.by}{Column name by which to group the data before calculating statistics.}

\item{...}{Additional arguments.}
}
\value{
If \code{add.Seurat} is TRUE, a Seurat object with appended metadata; otherwise, a data frame with the low-quality cell detection results.
}
\description{
This function handles multiple quality assessment methods to identify low-quality cells in single-cell data.
}
\details{
The 'Fail' value represents low-quality cells.
}
\examples{
\dontrun{
# Assuming 'seuratObj' is a Seurat object prepared for low-quality cell detection:
seuratObj <- RunLQ(object = seuratObj, methods = c("MAD", "ddqc", "miQC"), add.Seurat = TRUE)
}

}
\references{
1.Hippen, Ariel A et al. “miQC: An adaptive probabilistic framework for quality control of single-cell RNA-sequencing data.” PLoS computational biology vol. 17,8 e1009290. 24 Aug. 2021, doi:10.1371/journal.pcbi.1009290

2.Subramanian, Ayshwarya et al. “Biology-inspired data-driven quality control for scientific discovery in single-cell transcriptomics.” Genome biology vol. 23,1 267. 27 Dec. 2022, doi:10.1186/s13059-022-02820-w
}
\seealso{
\code{\link{RunLQ_MAD}}, \code{\link{RunLQ_miQC}}, \code{\link{RunLQ_ddqc}}
}
