% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunLQ_ddqc}
\alias{RunLQ_ddqc}
\title{Low quality cells detection with ddqc}
\usage{
RunLQ_ddqc(
  object,
  add.Seurat = TRUE,
  split.by = "orig.ident",
  BPtmpdir = "./temp/SingleCellMQC_ddqc/",
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{add.Seurat}{Logical; if TRUE, adds the ddqc results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the ddqc results.}

\item{split.by}{Column name by which to group the data before calculating statistics.}

\item{BPtmpdir}{Temporary directory for BPCells matrix processing.}

\item{...}{Additional arguments to be passed to ddqcR::ddqc.metrics function.}
}
\value{
A Seurat object with appended ddqc metadata if \code{add.Seurat} is TRUE; otherwise, a data frame with the ddqc results.
}
\description{
Performs quality control using the ddqc method on a Seurat object, an unsupervised adaptive QC framework.
}
\details{
The 'Fail' value represents low-quality cells.
}
\examples{
\dontrun{
# Assuming `seuratObject` is a Seurat object
seuratObject <- RunLQ_ddqc(seuratObject, add.Seurat = TRUE)
}
}
\references{
Subramanian, Ayshwarya et al. “Biology-inspired data-driven quality control for scientific discovery in single-cell transcriptomics.” Genome biology vol. 23,1 267. 27 Dec. 2022, doi:10.1186/s13059-022-02820-w
}
\seealso{
\code{\link{RunLQ}}
}
