% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunLQ_fixed}
\alias{RunLQ_fixed}
\title{Low quality cells detection based on fixed cutoff}
\usage{
RunLQ_fixed(
  object,
  min.nFeature_RNA = NULL,
  min.nCount_RNA = NULL,
  percent.mt = NULL,
  percent.rb = NULL,
  percent.hb = NULL,
  min.nFeature_ADT = NULL,
  min.nCount_ADT = NULL,
  add.Seurat = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object to be filtered.}

\item{min.nFeature_RNA}{(Optional) Minimum number of RNA features per cell.}

\item{min.nCount_RNA}{(Optional) Minimum RNA count per cell.}

\item{percent.mt}{(Optional) Maximum percentage of mitochondrial genes per cell.}

\item{percent.rb}{(Optional) Maximum percentage of ribosomal genes per cell.}

\item{percent.hb}{(Optional) Maximum percentage of hemoglobin genes per cell.}

\item{min.nFeature_ADT}{(Optional) Minimum number of ADT features per cell.}

\item{min.nCount_ADT}{(Optional) Minimum ADT count per cell.}

\item{add.Seurat}{(Optional) Logical. If TRUE, the final filtering result is added
to the Seurat object as a new metadata column named \code{db_fixed}.
Default is TRUE.}

\item{...}{Additional arguments.}
}
\value{
If \code{add.Seurat} is FALSE, returns a vector of "Pass" or "Fail" for each cell
based on the filtering criteria. If \code{add.Seurat} is TRUE, returns the modified
Seurat object with the final filtering result added as a new metadata column
named \code{db_fixed}.
}
\description{
Identifies low-quality cells based on the fixed cutoff on a Seurat object.
}
