% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunLQ_MAD}
\alias{RunLQ_MAD}
\title{Low quality cells detection with MAD}
\usage{
RunLQ_MAD(
  object,
  nmads = 3,
  split.by = NULL,
  add.Seurat = T,
  share.medians = F,
  share.mads = F,
  ...
)
}
\arguments{
\item{object}{Seurat object or data frame containing metrics for quality control. RNA metrics should contain "nCount_RNA", "nFeature_RNA" and "percent.mt". If you have ADT data, it should contain "nCount_ADT", "nFeature_ADT" and "percent.isotype". These metrics can all be generated via the \code{\link{CalculateMetrics}} function.}

\item{nmads}{Numeric. The number of MADs to use for identifying outliers. Default is 3.}

\item{split.by}{Optional; Column name by which to group the data before calculating statistics.}

\item{add.Seurat}{Logical; if TRUE, adds the MAD results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the MAD results.}

\item{share.medians}{Logical; if TRUE, forces all groups to have the same median. Default is FALSE.}

\item{share.mads}{Logical; if TRUE, forces all groups to have the same MAD. Default is FALSE.}

\item{...}{Additional arguments to be passed to \code{\link[scater]{isOutlier}} function from the \code{scater} package.}
}
\value{
If \code{add.Seurat} is TRUE, a Seurat object with appended metadata; otherwise, a data frame with outlier results.
}
\description{
Identifies low-quality cells based on the median absolute deviation (MAD) method on a Seurat object.
}
