% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_QC.R
\name{RunLQ_miQC}
\alias{RunLQ_miQC}
\title{Low quality cells detection with miQC}
\usage{
RunLQ_miQC(
  object,
  model_type = "linear",
  split.by = "orig.ident",
  name.nFeature_RNA = "nFeature_RNA",
  name.percent.mt = "percent.mt",
  add.Seurat = TRUE,
  random.seed = 1,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{model_type}{(character) What type of model to generate. A linear mixture model ("linear") based on mitochondrial percentage and library complexity is recommended. B-spline ("spline") and two-degree polynomial ("polynomial") models are also supported. For a simpler model, a one-dimensional gaussian mixture model ("one_dimensional") based on mitochondrial percentage only is available. Default = "linear".}

\item{split.by}{Column name by which to group the data before calculating statistics.}

\item{name.nFeature_RNA}{The name of the column in metadata representing the number of detected features per cell.}

\item{name.percent.mt}{The name of the column in metadata representing MT pct per cell.}

\item{add.Seurat}{Logical; if TRUE, adds the miQC results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the miQC results.}

\item{random.seed}{Random seed for reproducibility.}

\item{...}{Additional arguments to be passed to \code{\link[miQC]{filterCells}} function from the \code{miQC} package.}
}
\value{
A Seurat object with appended miQC metadata if \code{add.Seurat} is TRUE; otherwise, a data frame with the miQC results.
}
\description{
Performs quality control using the miQC method on a Seurat object, an adaptive probabilistic framework for quality control of single-cell RNA-sequencing data.
}
\details{
The 'Fail' value represents low-quality cells.
}
\examples{
\dontrun{
# Assuming `seuratObject` is a Seurat object
seuratObject <- RunLQ_miQC(seuratObject, add.Seurat = TRUE)
}
}
\references{
Hippen, Ariel A et al. “miQC: An adaptive probabilistic framework for quality control of single-cell RNA-sequencing data.” PLoS computational biology vol. 17,8 e1009290. 24 Aug. 2021, doi:10.1371/journal.pcbi.1009290
}
\seealso{
\code{\link{RunLQ}}
}
