% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowQuality.R
\name{RunLQ_VDJ}
\alias{RunLQ_VDJ}
\title{VDJ Low quality cells detection in single cell TCR and/or BCR data.}
\usage{
RunLQ_VDJ(object, add.Seurat = T, ...)
}
\arguments{
\item{object}{A single-cell object.}

\item{add.Seurat}{Logical; if TRUE, adds the detection results as metadata to the Seurat object and stored in \code{SingleCelMQC} slot in \code{misc}. Otherwise, returns a data.frame of the detection results.}

\item{...}{Additional arguments}
}
\value{
If \code{add.Seurat} is TRUE, returns the modified Seurat object. Otherwise, returns a list or data.frame.
}
\description{
Performs Low quality cells based on VDJ single chain.
}
\details{
The classification as 'Pass' or 'Fail'. 'Fail' represents Low quality cells (single chain).
}
