% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{RunPipeline}
\alias{RunPipeline}
\title{Initial Clustering of Single-Cell Data}
\usage{
RunPipeline(
  object,
  threads.pca = NULL,
  threads.knn_hnsw = 1,
  BPtmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  BP.cutoff = 2e+05,
  BPdir = "./BPCellData/",
  add.BPscale_data = F,
  preprocess = "rna.umap",
  resolution = 1,
  har.group.by = "orig.ident",
  dims = 1:20,
  wnn.upper.cutoff = 2e+05,
  wnn.knn.range = 300,
  wnn.prune.SNN = 1/20,
  har.seed = 1,
  nfeatures = 2000,
  exclude_patterns = c("^IGH[VDJ].*", "^IGK[JV].*", "^IGL[JLV].*", "^IGL[ON].*",
    "^TRA[JV].*", "^TRB[VDJ].*")
)
}
\arguments{
\item{object}{A Seurat object.}

\item{threads.pca}{Number of threads to use for PCA computation for BPCells. Default is NULL (automatic).}

\item{threads.knn_hnsw}{Number of threads to use for KNN/HNSW computation for BPCells. Default is 1.}

\item{BPtmpdir}{Temporary directory for BPCells processing. Default is "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/".}

\item{BP.cutoff}{The min cutoff for the number of cells to convert to BPCells processing data. Default is 200000.}

\item{BPdir}{Directory for BPCells data storage. Default is "./BPCellData/".}

\item{add.BPscale_data}{Add BPCells scale data. Default is FALSE.}

\item{preprocess}{The method of preprocessing and clustering to use. Options include:
\itemize{
\item "rna.pca": Perform the analysis to PCA step on RNA assay.
\item "rna.umap": Perform the analysis to UMAP step on RNA assay.
\item "rna.umap.har": Perform the analysis to UMAP step on RNA assay with Harmony integration.
\item "adt.pca": Perform the analysis to PCA step on ADT assay.
\item "adt.umap": Perform the analysis to UMAP step on ADT assay.
\item "adt.umap.har": Perform the analysis to UMAP step on ADT assay with Harmony integration.
\item "wnn.umap": Perform the analysis to UMAP step using weighted nearest neighbor (WNN) method on RNA+ADT assays.
\item "wnn.umap.har": Perform the analysis to UMAP step using weighted nearest neighbor (WNN) method on RNA+ADT assays with Harmony integration.
}
The default setting: "rna.umap".}

\item{resolution}{The resolution for clustering. Default is 1.}

\item{har.group.by}{The column name in meta.data to group by for harmony integration. Default is "orig.ident".}

\item{dims}{The dimensions to use for analysis. Default is 1:20.}

\item{wnn.upper.cutoff}{The max cutoff for the number of cells to perform WNN analysis. Default is 200000.}

\item{wnn.knn.range}{KNN range for WNN processing. Default is 300.}

\item{wnn.prune.SNN}{Pruning parameter for WNN SNN graph. Default is 1/20.}

\item{har.seed}{Seed for harmony integration. Default is 1.}

\item{nfeatures}{Number of features to select as top variable features. Default is 2000.}

\item{exclude_patterns}{Regular expression patterns to exclude from variable features. Default is:\verb{^IGH[VDJ]*}, \verb{^IGK[JV]*}, \verb{^IGL[JLV]*}, \verb{^IGL[ON].*}, \verb{^TRA[JV]*}, \verb{^TRB[VDJ]*}}
}
\value{
The Seurat object with preprocessing and clustering results included, such as PCA scores, UMAP coordinates, and cluster assignments.
}
\description{
Default single-cell data preprocessing, including normalization, feature selection, scaling, dimensionality reduction, and clustering.
}
\details{
For more details on these methods, please see the \code{\link{Seurat}} and BPCells documentation.
If the number of cells is greater than the \code{BP.cutoff}, the function will convert data to BPCells type.
If the number of cells is greater than the \code{wnn.upper.cutoff}, the function will perform WNN analysis.
}
\examples{
\dontrun{
##do rna umap analysis
seuratObject <- RunPipeline(seuratObject, preprocess = "rna.umap", resolution = 1)

##do adt pca analysis
seuratObject <- RunPipeline(seuratObject, preprocess = "adt.pca")

##do wnn umap analysis
seuratObject <- RunPipeline(seuratObject, preprocess = "wnn.umap", resolution = 1)

##do rna harmony umap analysis
seuratObject <- RunPipeline(seuratObject, preprocess = "rna.umap.har", resolution = 1)

}

}
