% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{RunPipelinePerGroup}
\alias{RunPipelinePerGroup}
\title{Run Single-Cell Analysis Pipeline on Each Group}
\usage{
RunPipelinePerGroup(object, split.by = "orig.ident", ...)
}
\arguments{
\item{object}{A Seurat object.}

\item{split.by}{A metadata column name used to split the Seurat object into separate groups (e.g., orig.ident).
Default is "orig.ident".}

\item{...}{Additional arguments passed to the \code{RunPipeline} function. See \code{RunPipeline} documentation for details.}
}
\value{
A list of Seurat objects, where each element corresponds to the result of the \code{RunPipeline} function
applied to a specific group.
}
\description{
This function splits a Seurat object into multiple groups based on a specified metadata column (e.g., orig.ident),
and then applies the RunPipeline function to each group independently. This is useful for batch-specific analysis
or when you want to preprocess and cluster each group separately.
}
