% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_generation.R
\name{RunPseudobulkData}
\alias{RunPseudobulkData}
\title{Create Pseudobulk Data}
\usage{
RunPseudobulkData(
  object,
  assay = "RNA",
  slot = "counts",
  sample.by = NULL,
  cluster.by = NULL,
  other_cols_contain = NULL,
  method = "sum"
)
}
\arguments{
\item{object}{A single-cell object (e.g., Seurat object) from which
expression data and metadata will be extracted.}

\item{assay}{A character string specifying the assay to use (e.g., "RNA").
Defaults to "RNA".}

\item{slot}{A character string specifying the slot to use for expression
data (e.g., "counts", "data"). Defaults to "counts".}

\item{sample.by}{A character string specifying the metadata column representing
the samples (e.g., patient ID, donor ID). This defines the primary unit
for pseudobulk.}

\item{cluster.by}{(Optional) A character string specifying the metadata column
representing cell clusters or cell types. If provided, pseudobulk matrices
will be generated separately for each unique cluster, resulting in a list
of matrices. Defaults to \code{NULL}.}

\item{other_cols_contain}{(Optional) A character vector of additional metadata
column names to include in the pseudobulk metadata. Defaults to \code{NULL}.}

\item{method}{A character string indicating the aggregation method for pseudobulking.
Accepted values depend on the underlying \code{BPCells::pseudobulk_matrix} function,
common options include "sum" (default), "mean".}
}
\value{
A list containing three elements:
\itemize{
\item \code{pseudobulk_mat}: A matrix or a named list of matrices of pseudobulk
expression values (genes x pseudobulk samples). If \code{cluster.by} is
specified, it's a list where names are clusters.
\item \code{pseudobulk_metadata}: A data frame of metadata for the pseudobulk
samples. Rows correspond to pseudobulk samples, and columns include
\code{sample.by} and any \code{other_cols_contain}.
\item \code{ncells}: A data frame summarizing the number of cells contributing
to each pseudobulk sample. Includes columns for \code{sample.by}, \code{cluster.by}
(if specified), and "ncells".
}
}
\description{
This function generates a pseudobulk object consisting of expression matrices,
corresponding metadata, and cell count information. It aggregates single-cell
expression data into pseudobulk samples based on specified grouping variables.
}
\details{
This function leverages \code{BPCells::pseudobulk_matrix} for efficient aggregation.
It constructs unique keys for aggregation from \code{sample.by} and \code{cluster.by}
(if provided). The pseudobulk metadata is constructed by aggregating unique or
mean values of specified metadata columns for each pseudobulk sample.
}
