% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_function.R
\name{RunReport}
\alias{RunReport}
\title{Generate QC HTML Report}
\usage{
RunReport(
  object = NULL,
  VDJ_data = NULL,
  sample.by = "orig.ident",
  outputFile = "./SingleCellMQC/SingleCellMQC.html",
  color = NULL,
  section.sample = c(1:3),
  section.cell = c(1:2),
  section.feature = 1,
  section.batch = c(1:3),
  RNA_cluster_name = "rna_cluster",
  ADT_cluster_name = "adt_cluster",
  RNA.batch.by = "orig.ident",
  ADT.batch.by = "orig.ident",
  RNA.batch.covariate = c("orig.ident", "percent.mt", "nFeature_RNA", "nCount_RNA"),
  ADT.batch.covariate = c("orig.ident", "nFeature_ADT", "nCount_ADT"),
  celltype.by = "ScType",
  tissue = NULL
)
}
\arguments{
\item{object}{Seurat object containing single-cell data. This object should be preprocessed using the \code{RunPreprocess} function or other appropriate preprocessing functions.}

\item{VDJ_data}{List of VDJ data. Default is NULL.}

\item{sample.by}{Column name in the metadata indicating sample information. Default is "orig.ident".}

\item{outputFile}{Path to the output HTML file. Default is "./SingleCellMQC/SingleCellMQC.html".}

\item{color}{Color palette for plots. Default is NULL.}

\item{section.sample}{Sections to include in the Sample QC module. Default is \code{c(1:3)}.
\itemize{
\item \strong{\code{1}}: Run Sample QC: Sample quality assessment.
\item \strong{\code{2}}: Run Sample QC: Outlier sample detection by cell type.
\item \strong{\code{3}}: Run Sample QC: Sample identity verification.
}}

\item{section.cell}{Sections to include in the Cell QC module. Default is \code{c(1:2)}.
\itemize{
\item \strong{\code{1}}: Run Cell QC: Low-quality cell information.
\item \strong{\code{2}}: Run Cell QC: Doublet information.
}}

\item{section.feature}{Sections to include in the Feature QC module. Default is \code{1}.
\itemize{
\item \strong{\code{1}}: Run Feature QC: Feature quality assessment.
}}

\item{section.batch}{Sections to include in the Batch QC module. Default is \code{1:3}.
\itemize{
\item \strong{\code{1}}: Run Batch QC: Sample level.
\item \strong{\code{2}}: Run Batch QC: Cell level.
\item \strong{\code{3}}: Run Batch QC: Feature level.
}}

\item{RNA_cluster_name}{Column name in the metadata indicating RNA cluster information. Default is "rna_cluster". Only for: section.batch (2).}

\item{ADT_cluster_name}{Column name in the metadata indicating ADT cluster information. Default is "adt_cluster". Only for: section.batch (2).}

\item{RNA.batch.by}{Character vector. Column name(s) in the metadata indicating RNA batch information (categorical variable) or other categorical variables for PCA or UMAP plot. Default is "orig.ident". Only for: section.batch (1, 2).}

\item{ADT.batch.by}{Character vector. Column name(s) in the metadata indicating ADT batch information (categorical variable) or other categorical variables for PCA or UMAP plot. Default is "orig.ident". Only for: section.batch (1, 2).}

\item{RNA.batch.covariate}{Character vector. Covariates (can be categorical or continuous) for RNA \code{SVD} and \verb{Variance Explained} analysis. Default is \code{c("orig.ident", "percent.mt", "nFeature_RNA", "nCount_RNA")}. section.batch (1, 3).}

\item{ADT.batch.covariate}{Character vector. Covariates (can be categorical or continuous) for ADT \code{SVD} and \verb{Variance Explained} analysis. Default is \code{c("orig.ident", "nFeature_ADT", "nCount_ADT")}. Only for: section.batch (1, 3).}

\item{celltype.by}{Column name in the metadata indicating cell type information. Default is "ScType". Only for: section.sample (2), section.batch (1).}

\item{tissue}{Tissue name. Default is NULL. Only for: section.sample (1, 2).}
}
\value{
A HTML report is generated and saved to the specified output file path. No value is returned.
}
\description{
This function generates a comprehensive HTML report for single-cell data quality control (QC).
The report includes modules for sample QC, cell QC, feature QC, and batch QC, with customizable sections for each module.
}
