% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{RunScType}
\alias{RunScType}
\title{Run Cell Annotation with ScType}
\usage{
RunScType(
  object,
  split.by = NULL,
  return.name = "ScType",
  genelist = NULL,
  use_internal_data = TRUE,
  group.by = "seurat_clusters",
  preprocess = NULL,
  resolution = 1,
  cutoff = 3,
  type.tissue = NULL,
  type.condition = NULL,
  type.cell = NULL,
  data_source = "Main",
  ntop = 30,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{split.by}{A character string specifying the metadata column to split the data by before prediction. Default: NULL.}

\item{return.name}{A character string specifying the name of the new column to add to the Seurat object. Default: "ScType".}

\item{genelist}{A list of gene lists to use for prediction. Default: NULL.}

\item{use_internal_data}{Logical; if TRUE, uses an internal sctype dataset of immune cell markers for prediction. Default: TRUE.}

\item{group.by}{A character string specifying the metadata column to group cells by before prediction. Default: "seurat_clusters".}

\item{preprocess}{A character string specifying the dimensionality reduction method. Supported values include:
\itemize{
\item "rna.umap": Perform the analysis to UMAP step on RNA assay.
\item "adt.umap": Perform the analysis to UMAP step on ADT assay.
\item "wnn.umap": Perform the analysis to UMAP step using weighted nearest neighbor (WNN) method on RNA+ADT assays.
}}

\item{resolution}{The resolution to use for clustering. Default is 1.}

\item{cutoff}{The cutoff value used for deciding when to label a cell type as 'Unknown'. Default:4}

\item{type.tissue}{A character string specifying the tissue type to use for marker selection. Default: "Immune system".}

\item{type.condition}{A character string specifying the condition type to use for marker selection. Default: NULL.}

\item{type.cell}{A character string specifying the cell type to use for marker selection. Default: "Normal".}

\item{data_source}{A character string specifying the data source to use for marker selection. Default: "ScType".
\itemize{
\item "ScType": Use the ScType database for marker selection.
\item "Cell_Taxonomy": Use the Cell_Taxonomy database for marker selection.
}}

\item{ntop}{An integer specifying the number of top markers to use for prediction. Default: 30.}

\item{...}{Additional arguments to pass to the \code{RunPipeline} function.}
}
\value{
A Seurat object with a new column "ScType" in the metadata slot. The column contains the predicted cell types.
}
\description{
This function perfroms cell annotation using ScType method on a Seurat object.
}
\details{
The function first preprocesses the data if requested, and then uses the provided or internal gene list to score cells based on the ScType method.
}
\examples{
\dontrun{
# Assuming `seuratObj` is a Seurat object
seuratObj <- RunScType(seuratObj, group.by="seurat_clusters", type.tissue = "Immune system")
}

}
\references{
Ianevski, Aleksandr et al. “Fully-automated and ultra-fast cell-type identification using specific marker combinations from single-cell transcriptomic data.” Nature communications vol. 13,1 1246. 10 Mar. 2022, doi:10.1038/s41467-022-28803-w
}
\seealso{
\code{\link{ShowDatabaseTissue}} for displaying the available tissue types in the ScType database.
}
