% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_effect.R
\name{RunVarExplainedPerFeature}
\alias{RunVarExplainedPerFeature}
\title{Calculate Per Molecular Feature Variance Explained by Variables}
\usage{
RunVarExplainedPerFeature(
  object,
  type = "cell",
  assay = "RNA",
  pseudobulk.sample.by = "orig.ident",
  variables = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{type}{Analysis type, must be either "cell" (single-cell level) or "pseudobulk" (pseudobulk level)}

\item{assay}{Name of assay to use, default is "RNA"}

\item{pseudobulk.sample.by}{Metadata sample column name to group cells for pseudobulk analysis (when type="pseudobulk")}

\item{variables}{Variables to calculate variance explained for (must be specified)}
}
\value{
Return variance explained results for each molecular feature
}
\description{
This function calculates the percentage of variance explained by one or more variables of interest for each molecular feature (e.g., gene, protein).
}
