% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariancePartition.R
\name{RunVarPartMetadata}
\alias{RunVarPartMetadata}
\title{Run Variance Partitioning on Cell Metadata}
\usage{
RunVarPartMetadata(object, features, formula, split.by = NULL, do.log1p = T)
}
\arguments{
\item{object}{An object containing cell metadata.}

\item{features}{A character vector specifying the names of the metadata
features to be analyzed. Common examples include "percent.mt", "percent.rb",
"percent.hb", "percent.dissociation".}

\item{formula}{A \code{formula} object specifying the fixed and random effects
to be included in the linear mixed model. This formula should only contain
the explanatory variables (e.g., \code{~ (1|Batch)}). The
response variable will be dynamically added within the function for each
\code{feature}.}

\item{split.by}{(Optional) A character string specifying the name of a
column in the metadata by which to split the data for separate variance
partition analyses. If \code{NULL} (default), the analysis is performed on
the entire object.}

\item{do.log1p}{A logical value indicating whether to apply a \code{log1p}
transformation (i.e., \code{log(1 + x)}) to the specified \code{features} before
performing the variance partition. Defaults to \code{TRUE}.}
}
\value{
A data frame summarizing the variance explained by each
component in the \code{formula} for each \code{feature} and (if \code{split.by} is used)
for each group. The data frame includes columns for 'Group', 'Feature',
and columns for each variance component.
}
\description{
This function performs variance partition analysis on specified cell metadata
features using linear mixed models. It can analyze the entire dataset or
split it based on a grouping variable.
}
\details{
This function leverages \code{lmerTest::lmer} to fit linear mixed models and
\code{variancePartition::calcVarPart} to calculate the fraction of variance
explained by each covariate. If \code{split.by} is provided, the function
iterates through each unique value in the specified column, performs the
variance partition for that subset of data, and combines the results.
}
