% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariancePartition.R
\name{RunVarPartPseudobulk}
\alias{RunVarPartPseudobulk}
\title{Run Variance Partitioning on Pseudobulk Data}
\usage{
RunVarPartPseudobulk(object, variables = NULL, formula = NULL)
}
\arguments{
\item{object}{A list object containing pseudobulk data gets from \code{RunPseudobulkData} function.}

\item{variables}{(Optional) A character vector of column names from
\code{pseudobulk_metadata} to be used as random effects in the variance
partition model. If provided, the formula will be constructed using \code{(1|variable)}.
Cannot be used with \code{formula}.}

\item{formula}{(Optional) A \code{formula} object specifying the fixed and random
effects for the variance partition model (e.g., \code{~ (1|donor_id)}).
If provided, \code{variables} is ignored. If neither \code{variables} nor \code{formula}
are provided, an error will be thrown.}
}
\value{
A data frame summarizing the variance explained by each
component in the \code{formula} for each gene. The data frame includes columns
for 'Group' (indicating if multiple groups were processed), 'Feature' (gene name),
and columns for each variance component.
}
\description{
This function performs variance partitioning analysis on pseudobulk
data. It can analyze either a single pseudobulk matrix or a list of matrices
(e.g., grouped by cell type), using \code{dreamlet} and \code{variancePartition} packages.
}
\details{
This function internally uses \code{dreamlet::processOneAssay} to prepare the data
and \code{variancePartition::fitExtractVarPartModel} to perform the actual
variance partitioning.
If \code{pseudobulk_mat} is a list, the function iterates through each element,
performing variance partitioning independently for each group/cluster.
}
\seealso{
\code{\link[dreamlet]{processOneAssay}}, \code{\link[variancePartition]{fitExtractVarPartModel}}
}
