% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariancePartition.R
\name{RunVarPartPseudobulkPCA}
\alias{RunVarPartPseudobulkPCA}
\title{Run Variance Partitioning on Pseudobulk PCA embeddings}
\usage{
RunVarPartPseudobulkPCA(object, variables = NULL, nPCs = 2, formula = NULL)
}
\arguments{
\item{object}{A list object containing pseudobulk data gets from \code{RunPseudobulkData} function.}

\item{variables}{(Optional) A character vector of column names from
\code{pseudobulk_metadata} to be used as random effects in the variance
partition model. If provided, the formula will be constructed using \code{(1|variable)}.
Cannot be used with \code{formula}.}

\item{nPCs}{An integer specifying the number of top principal components
to analyze. Defaults to 2. The actual number of PCs analyzed will be
the minimum of \code{nPCs} and the available number of PCs.}

\item{formula}{(Optional) A \code{formula} object specifying the fixed and random
effects for the variance partition model (e.g., \code{~  (1|donor_id)}).
If provided, \code{variables} is ignored. If neither \code{variables} nor \code{formula}
are provided, an error will be thrown.}
}
\value{
A data frame summarizing the variance explained by each
component in the \code{formula} for each PC. The data frame includes columns
for 'Group' (indicating if multiple groups were processed), 'Feature' (PC name, e.g., 'PC_1'),
and columns for each variance component.
}
\description{
This function performs variance partitioning analysis on the Principal Component
(PC) embeddings derived from pseudobulk expression data. It allows identifying
the factors contributing to the variance in the PCs.
}
\details{
This function first performs PCA on the pseudobulk expression data using
an internal \code{runPseudobulkPCA} function (assumed to return a Seurat-like
object). It then extracts the specified number of top PCs and treats each
PC's projection as a "feature" for a linear mixed model.
\code{lmerTest::lmer} is used to fit these models, and \code{variancePartition::calcVarPart}
calculates the variance explained by each covariate. If \code{pseudobulk_mat} is a
list, PCA and variance partitioning are performed independently for each
group/cluster.
}
