% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_function.R
\name{ShowCellMetricsName}
\alias{ShowCellMetricsName}
\title{Show available cell quality control metrics}
\usage{
ShowCellMetricsName(object)
}
\arguments{
\item{object}{A Seurat object containing the per-cell quality control metrics. The input must be a valid Seurat object to extract the quality control data.}
}
\value{
A character vector containing the names of the available cell quality control metrics in the Seurat object.
}
\description{
This function retrieves and displays the names of available cell-level quality control metrics from a Seurat object.
}
