% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{ShowCommonPCTTissue}
\alias{ShowCommonPCTTissue}
\title{Show Common PCT Tissue}
\usage{
ShowCommonPCTTissue(object = NULL)
}
\arguments{
\item{object}{NULL}
}
\value{
A character vector containing the available tissue types.
}
\description{
Show Common PCT Tissue
}
