% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{ShowDatabaseTissue}
\alias{ShowDatabaseTissue}
\title{Show Database Tissue}
\usage{
ShowDatabaseTissue(database = "ScType")
}
\arguments{
\item{database}{A character string specifying the database to use. Supported values include:
\itemize{
\item "ScType": Use the ScType database.
\item "Cell_Taxonomy": Use the Cell_Taxonomy database.
}}
}
\value{
A character vector of unique tissue types.
}
\description{
This function returns the unique tissue types in the ScType or Cell_Taxonomy database.
}
\examples{
ShowDatabaseTissue("ScType")
ShowDatabaseTissue("Cell_Taxonomy")

}
