% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSample.R
\name{ShowSampleMetricsName}
\alias{ShowSampleMetricsName}
\title{Show Sample Metrics}
\usage{
ShowSampleMetricsName(object, type = "count")
}
\arguments{
\item{object}{A Seurat object.}

\item{type}{The type of metrics to show. Supported values include "count", "summary", and "Metrics_10x". Default is "count".}
}
\value{
A character vector containing the available metrics for the samples in the Seurat object.
}
\description{
This function show the available metrics for the samples in the Seurat object.
}
\examples{
\dontrun{
# Assuming 'seurat_obj' is a Seurat object:
# Show the available sample count metrics for the samples in the Seurat object
ShowSampleMetrics(seurat_obj, type = "count")
}

}
