% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{StatsCellFilter}
\alias{StatsCellFilter}
\title{Generate Statistics for Cell Filtering}
\usage{
StatsCellFilter(
  object,
  type.detection = c("lq", "db"),
  return.type = c("table", "interactive_table")
)
}
\arguments{
\item{object}{A Seurat object or a data frame containing cell metadata.}

\item{type.detection}{Type of detection to analyze. Options are "lq" (low-quality cells) or "db" (doublets). Default is "lq".}

\item{return.type}{Type of output to return. Options are "table" (data frame) or "interactive_table" (interactive HTML table). Default is "table".}
}
\value{
A data frame or an interactive table containing the following columns:
\itemize{
\item \code{Method}: The filtering method used.
\item \code{Pre}: Number of cells before filtering.
\item \code{Post}: Number of cells after filtering.
\item \verb{Post\%}: Percentage of cells remaining after filtering.
\item \code{Filtered}: Number of cells filtered out.
\item \verb{Filtered\%}: Percentage of cells filtered out.
}
}
\description{
This function generates statistics for cell filtering based on low-quality (lq) or doublet (db) detection.
It calculates the number of cells before and after filtering, as well as the percentage of cells filtered out.
The results can be returned as a data table or an interactive table.
}
