% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{StatsFeaturePCT}
\alias{StatsFeaturePCT}
\title{Calculate Feature Percentage Across Groups}
\usage{
StatsFeaturePCT(
  object,
  feature = NULL,
  group.by = NULL,
  split.by = "orig.ident",
  assay = "RNA",
  slot = "counts",
  return.type = "table"
)
}
\arguments{
\item{object}{A Seurat object containing single-cell data.}

\item{feature}{A character vector specifying the features (e.g., genes) to analyze. Must be present in the Seurat object.}

\item{group.by}{A character string specifying the column in the metadata to group cells by. Default is NULL.}

\item{split.by}{A character string specifying the column in the metadata to split cells by. Default is "orig.ident".}

\item{assay}{The assay to use for feature extraction. Default is "RNA".}

\item{slot}{The slot to use for feature extraction. Default is "counts".}

\item{return.type}{Type of output to return. Options are "table" (data frame) or "interactive_table" (interactive HTML table). Default is "table".}
}
\value{
A data frame or an interactive table containing the following columns:
\itemize{
\item \code{Feature}: The feature (e.g., gene) analyzed.
\item \code{Group}: The group or split category (if specified).
\item \code{PCT}: The percentage of cells expressing the feature in the group.
}
}
\description{
This function calculates the percentage of cells expressing a given feature (e.g., a gene) across specified groups.
The results can be returned as a data table or an interactive table.
}
