% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{StepBPPCAToUMAP}
\alias{StepBPPCAToUMAP}
\title{Perform UMAP on PCA results using BPCells}
\usage{
StepBPPCAToUMAP(
  object,
  pca_name = "rna.pca",
  assay = "RNA",
  output_umap_name = "rna.umap",
  dims = 1:20,
  ...
)
}
\arguments{
\item{object}{A Seurat object or PCA result.}

\item{pca_name}{Name of the PCA object (default: "rna.pca").}

\item{assay}{The Seurat assay to use (default: "RNA").}

\item{output_umap_name}{Name of the UMAP output (default: "rna.umap").}

\item{dims}{Dimensions to use for UMAP (default: 1:20).}

\item{...}{Additional arguments passed to uwot::umap.}
}
\value{
A Seurat object with UMAP results or a UMAP matrix.
}
\description{
Perform UMAP on PCA results using BPCells.
}
