% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipeline.R
\name{StepBPToPCA}
\alias{StepBPToPCA}
\title{Perform preprocessing and PCA using BPCells}
\usage{
StepBPToPCA(
  object,
  assay = "RNA",
  nfeatures = 2000,
  exclude_patterns = NULL,
  threads = NULL,
  tmpdir = "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/",
  output_pca_name = "rna.pca",
  add.norm_data = TRUE,
  add.scale_data = FALSE
)
}
\arguments{
\item{object}{A Seurat object or BPCells matrix.}

\item{assay}{The Seurat object assay to use (default: "RNA").}

\item{nfeatures}{Number of features to select (default: 2000).}

\item{exclude_patterns}{Patterns to exclude from feature selection (default: NULL).}

\item{threads}{Number of threads to use (default: NULL).}

\item{tmpdir}{Temporary directory for intermediate files (default: "./temp/SingleCellMQC_BPCellsStepBPToPCAScale/").}

\item{output_pca_name}{Name of the PCA output (default: "rna.pca").}

\item{add.norm_data}{Whether to add normalized data to the Seurat object (default: TRUE).}

\item{add.scale_data}{Whether to add scaled data to the Seurat object (default: FALSE).}
}
\value{
A Seurat object with PCA results or a PCA matrix.
}
\description{
Perform preprocessing and PCA on a Seurat or BP matrix using BPCells.
}
