% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetrics}
\alias{CalculateMetrics}
\title{Calculate Multi-omics Metrics}
\usage{
CalculateMetrics(object, types = c("Sample", "Cell"), sample.by = "orig.ident")
}
\arguments{
\item{object}{A Seurat object or a list containing GEX (Seurat object), TCR (list; optional), and BCR (list; optional) components.}

\item{types}{A character vector specifying the types of metrics to run: "Sample", "Cell", and/or "Feature". Defaults to c("Sample", "Cell").}

\item{sample.by}{Character string specifying the grouping variable for sample metrics. Defaults to "orig.ident".}
}
\value{
A Seurat object. All results are stored in \code{meta.data} and \code{SingleCelMQC} slot in \code{misc}.
}
\description{
Calculates different metrics, including sample, cell, and feature metrics.
Utilizes \code{CalculateMetricsPerCell}, \code{CalculateMetricsPerSample}, and \code{CalculateMetricsPerFeature} to perform its tasks based on the \code{types} argument.
}
\seealso{
\code{\link{CalculateMetricsPerCell}}, \code{\link{CalculateMetricsPerSample}}, \code{\link{CalculateMetricsPerFeature}}
}
