% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{CalculateMetricsPerCell}
\alias{CalculateMetricsPerCell}
\title{Calculate Cell QC Metrics}
\usage{
CalculateMetricsPerCell(object)
}
\arguments{
\item{object}{The object is obtained by the \code{Read10XData} function, or a list containing GEX (Seurat object), TCR (list; optional), and BCR (list; optional) components, or a Seurat object.}
}
\value{
A Seurat object. All results are stored in \code{meta.data} and \code{SingleCelMQC} slot in \code{misc}
}
\description{
A generic function for Calculating cell-specific metrics based on the data type.
}
