% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_data_operations.R
\name{ConvertToSCE}
\alias{ConvertToSCE}
\alias{ConvertToSCE.Seurat}
\alias{ConvertToSCE.SingleCellExperiment}
\alias{ConvertToSCE.default}
\title{Convert an object to a SingleCellExperiment}
\usage{
ConvertToSCE(object, ...)

\method{ConvertToSCE}{Seurat}(object, assay = "RNA", slot = "counts", ...)

\method{ConvertToSCE}{SingleCellExperiment}(object, ...)

\method{ConvertToSCE}{default}(object, ...)
}
\arguments{
\item{object}{The input object to convert. Can be a Seurat object, SingleCellExperiment
object, or a matrix-like object.}

\item{...}{Additional arguments passed to the conversion method.}

\item{assay}{The name of the assay to extract from the Seurat object (default: "RNA").
Only applicable when \code{object} is a Seurat object.}

\item{slot}{The slot to extract from the Seurat object (default: "counts").
Only applicable when \code{object} is a Seurat object.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function converts an input object (e.g., Seurat, matrix-like, or SingleCellExperiment)
into a SingleCellExperiment object. If the input is already a SingleCellExperiment,
it is returned as-is. For Seurat objects, the specified assay and slot are used for
conversion. For other objects (e.g., matrices), they are wrapped into a
SingleCellExperiment object.
}
