% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_QC.R
\name{FindCommonPCTOutlier}
\alias{FindCommonPCTOutlier}
\title{Find Common PCT Outliers}
\usage{
FindCommonPCTOutlier(
  object,
  sample.by = "orig.ident",
  celltype.by = "ScType",
  return.type = c("table"),
  tissue = NULL
)
}
\arguments{
\item{object}{A Seurat object or a data frame containing the metadata.}

\item{sample.by}{A character string specifying the metadata column to use for identifying samples. Default is "orig.ident".}

\item{celltype.by}{A character string specifying the metadata column to use for identifying cell types. Default is "ScType".}

\item{return.type}{A character vector specifying the type of output to return. Options are "table" and/or "interactive_table". Default is "table".}

\item{tissue}{A character string specifying the tissue type to filter by. If NULL, the function will stop and prompt the user to specify a tissue.}
}
\value{
Depending on the \code{return.type} parameter, the function returns:
\itemize{
\item If "table" is specified, a data frame containing the outlier results.
\item If "interactive_table" is specified, an interactive table generated using the \code{reactable} package.
\item If both are specified, a list containing both the table and the interactive table.
}
}
\description{
This function identifies common pct (percentage) outliers based on a built-in database of cell proportions.
}
\details{
To detect outlier samples based on cell type composition, one approach involves compiling and integrating data from established tissue atlases,
such as those from the DISCO database and additional single-cell reference datasets,
to define the common range of inter-individual variability for major typical cell types within each tissue.
The function then identifies samples exhibiting deviations from these established reference proportions, marking them as potential outliers.
}
