% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genelist_Benchmark.R
\name{GetNoiseGeneList}
\alias{GetNoiseGeneList}
\title{Get Predefined Noise Gene Lists}
\usage{
GetNoiseGeneList(object)
}
\arguments{
\item{object}{A single-cell object (e.g., Seurat object) from which gene names
will be extracted.}
}
\value{
A named list, where each element is a character vector of gene names
belonging to a specific noise category:
\itemize{
\item \code{mt}: Mitochondrial genes (e.g., "MT-").
\item \code{rb}: Ribosomal protein genes (e.g., "RPS", "RPL").
\item \code{hb}: Hemoglobin genes.
\item \code{dissociation}: Genes commonly associated with dissociation stress.
\item \code{TCRab}: T-cell receptor alpha/beta chain V(D)J genes.
\item \code{BCR}: B-cell receptor (immunoglobulin, V(D)J) genes.
\item \code{Sex_chromosome}: Genes located on X and Y chromosomes.
}
Genes lists are intersected with the actual gene names present in the \code{object}.
}
\description{
This function generates a named list of common "noise" gene sets that are
often excluded or analyzed separately in single-cell RNA-seq data.
}
\details{
The function contains predefined patterns and lists of genes for various
noise categories. It matches these against the rownames of the expression
matrix from the provided \code{object} to return only the genes present in the dataset.
}
