% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFeature.R
\name{PlotFeatureMetrics}
\alias{PlotFeatureMetrics}
\title{Plot Feature Metrics}
\usage{
PlotFeatureMetrics(
  object,
  assay = "RNA",
  metric = "pct",
  sample = NULL,
  log.x = F,
  color = "#0072B2",
  ntop = 10,
  ylab = metric
)
}
\arguments{
\item{object}{A Seurat object processed through the \code{\link{CalculateMetrics}} or \code{\link{CalculateMetricsPerFeature}} function.}

\item{assay}{A character string specifying the assay to analyze. The default is \code{"RNA"}.}

\item{metric}{A character string specifying the feature metric to visualize. Default is "pct".}

\item{sample}{Optional. Character string specifying a sample name to subset the data. If \code{NULL}, the entire dataset is used.}

\item{log.x}{Logical. If \code{TRUE}, the x-axis will be scaled logarithmically. Default is \code{TRUE}.}

\item{color}{Character string specifying the color of the plot (in hex or other color formats). Default is \code{"#0072B2"}.}

\item{ntop}{Numeric. Specifies the number of top features to highlight in the bar plot. Default is \code{10}.}

\item{ylab}{Character string specifying the y-axis label for the top bar plot. Default is the same as the \code{metric} parameter.}
}
\value{
The function returns a \code{ggplot} object.
}
\description{
This function generates plots for visualizing the distribution of specific feature metrics.
}
