% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_effect.R
\name{PlotReducedDim}
\alias{PlotReducedDim}
\title{Visualize Reduced Dimensional Data}
\usage{
PlotReducedDim(
  object,
  group.by = "seurat_clusters",
  plot.type = "cell",
  reduction = "rna.umap",
  split.by = NULL,
  ncol = NULL,
  ggside = F,
  color = NULL,
  guide.nrow = 10,
  raster.cutoff = 1e+05,
  do.preprocess = F,
  preprocess = "rna.umap",
  sample.by.pseudobulk = "orig.ident",
  group.by.pseudobulk = "orig.ident",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing the single-cell data to be analyzed.}

\item{group.by}{A character string specifying the metadata column used for grouping cells (default: \code{"seurat_clusters"}).}

\item{plot.type}{A character string specifying the type of plot to generate. Options include \code{"cell"} and \code{"pseudobulk"} (default: \code{"cell"}).}

\item{reduction}{A character string specifying the type of reduction to plot (default: \code{"rna.umap"}).}

\item{split.by}{A character string specifying the metadata column used to split the data into subsets (default: \code{NULL}).}

\item{ncol}{An integer specifying the number of columns for the facet plot (default: \code{NULL}).}

\item{ggside}{Logical, whether to use ggside for facet plots (default: \code{FALSE}).}

\item{color}{A vector of colors to use for visualizations (default: \code{NULL}).}

\item{guide.nrow}{An integer specifying the number of rows for the legend (default: \code{10}).}

\item{raster.cutoff}{An integer specifying the cutoff for rasterization (default: \code{100000}).}

\item{do.preprocess}{Logical, whether to preprocess the data before plotting (default: \code{FALSE}).}

\item{preprocess}{A character string specifying the type of preprocessing to apply (default: \code{"rna.umap"}).}

\item{sample.by.pseudobulk}{A character string specifying the metadata column used for defining samples in the pseudobulk analysis (default: \code{"orig.ident"}).}

\item{group.by.pseudobulk}{A character string specifying the metadata column used for grouping cells in the pseudobulk analysis (default: \code{"orig.ident"}).}

\item{...}{Additional arguments to be passed to the plotting functions.}
}
\value{
A plot based on the specified \code{plot.type}.
}
\description{
This function generates scatter plots of reduced dimensional data, either at the cell level or pseudobulk level. It supports two main visualization types: cell-level scatter plots (\code{"cell"}) and pseudobulk PCA plots (\code{"pseudobulk"}).
}
\examples{
\dontrun{
# Visualize reduced dimensional data
PlotReducedDim(object, group.by = "seurat_clusters", plot.type = "cell")
}
}
\seealso{
\code{\link{PlotSamplePCA}}
}
