% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_new.R
\name{PlotReducedDim}
\alias{PlotReducedDim}
\title{Visualize Reduced Dimensional Data}
\usage{
PlotReducedDim(
  object,
  group.by = "seurat_clusters",
  process.pseudobulk = "pca",
  reduction = "rna.umap",
  split.by = NULL,
  ncol = NULL,
  ggside = F,
  color = NULL,
  guide.nrow = 10,
  raster.cutoff = 1e+05,
  size = 0.3
)
}
\arguments{
\item{object}{A Seurat object or a list of pseudobulk data.}

\item{group.by}{A character string specifying the metadata column used for grouping cells (default: \code{"seurat_clusters"}).}

\item{reduction}{A character string specifying the type of reduction to plot (default: \code{"rna.umap"}).}

\item{split.by}{A character string specifying the metadata column used to split the data into subsets (default: \code{NULL}).}

\item{ncol}{An integer specifying the number of columns for the facet plot (default: \code{NULL}).}

\item{ggside}{Logical, whether to use ggside for facet plots (default: \code{FALSE}).}

\item{color}{A vector of colors to use for visualizations (default: \code{NULL}).}

\item{guide.nrow}{An integer specifying the number of rows for the legend (default: \code{10}).}

\item{raster.cutoff}{An integer specifying the cutoff for rasterization (default: \code{100000}).}
}
\value{
A plot based on the specified \code{plot.type}.
}
\description{
This function generates scatter plots of reduced dimensional data, either at the cell level or pseudobulk level.
}
