% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_QC.R
\name{RunCorrection_DecontX}
\alias{RunCorrection_DecontX}
\title{Perform the contamination correction by DecontX}
\usage{
RunCorrection_DecontX(object, split.by = "orig.ident", ...)
}
\arguments{
\item{object}{A Seurat object.}

\item{split.by}{A character string specifying the metadata column in the Seurat object to split the data into batches.
Default is "orig.ident".}

\item{...}{Additional arguments passed to \code{decontX::decontX} for more customized analysis.}
}
\value{
A Seurat object contains additional corrected assay.
}
\description{
Perform the contamination correction by DecontX
}
