% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_function.R
\name{RunReport}
\alias{RunReport}
\title{Generate QC HTML Report}
\usage{
RunReport(
  object = NULL,
  VDJ_data = NULL,
  sample.by = "orig.ident",
  outputFile = "./SingleCellMQC/SingleCellMQC.html",
  color = NULL,
  section.sample = c(1:3),
  section.cell = c(1:2),
  section.feature = 1,
  section.batch = 1:2,
  RNA_cluster_name = "rna_cluster",
  ADT_cluster_name = "adt_cluster",
  RNA.batch.by = "orig.ident",
  ADT.batch.by = "orig.ident",
  RNA.batch.covariate = c("orig.ident", "percent.mt", "nFeature_RNA", "nCount_RNA"),
  ADT.batch.covariate = c("orig.ident", "nFeature_ADT", "nCount_ADT"),
  celltype.by = "ScType",
  tissue = NULL
)
}
\arguments{
\item{object}{Seurat object containing single-cell data. This object should be preprocessed using the \code{RunPreprocess} function or other appropriate preprocessing functions.}

\item{VDJ_data}{List of VDJ data. Default is NULL.}

\item{sample.by}{Column name in the metadata indicating sample information. Default is "orig.ident".}

\item{outputFile}{Path to the output HTML file. Default is "./SingleCellMQC/SingleCellMQC.html".}

\item{color}{Color palette for plots. Default is NULL.}

\item{section.sample}{Sections to include in the Sample QC module. Default is \code{c(1:3)}.
\itemize{
\item \strong{\code{1}}: Run Sample QC: Sample quality assessment.
\item \strong{\code{2}}: Run Sample QC: Outlier sample detection by cell type.
\item \strong{\code{3}}: Run Sample QC: Sample identity verification.
}}

\item{section.cell}{Sections to include in the Cell QC module. Default is \code{c(1:2)}.
\itemize{
\item \strong{\code{1}}: Run Cell QC: Low-quality cell information.
\item \strong{\code{2}}: Run Cell QC: Doublet information.
}}

\item{section.feature}{Sections to include in the Feature QC module. Default is \code{1}.
\itemize{
\item \strong{\code{1}}: Run Feature QC: Feature quality assessment.
}}

\item{section.batch}{Sections to include in the Batch QC module. Default is \code{1:2}.
\itemize{
\item \strong{\code{1}}: Run Batch QC: Clustering QC.
\item \strong{\code{2}}: Run Batch QC: Covariate Impact.
}}

\item{RNA_cluster_name}{Column name in the metadata indicating RNA cluster information. Default is "rna_cluster". Only for: \strong{Batch QC (Clustering QC)}.}

\item{ADT_cluster_name}{Column name in the metadata indicating ADT cluster information. Default is "adt_cluster". Only for: \strong{Batch QC (Clustering QC)}.}

\item{RNA.batch.by}{Column name in the metadata indicating RNA batch information. Default is "orig.ident". Only for: \strong{Batch QC (Clustering QC, Covariate Impact)}.}

\item{ADT.batch.by}{Column name in the metadata indicating ADT batch information. Default is "orig.ident". Only for: \strong{Batch QC (Clustering QC, Covariate Impact)}.}

\item{RNA.batch.covariate}{Covariates for RNA batch analysis. Default is \code{c("orig.ident", "percent.mt", "nFeature_RNA", "nCount_RNA")}. Only for: \strong{Batch QC (Covariate Impact)}.}

\item{ADT.batch.covariate}{Covariates for ADT batch analysis. Default is \code{c("orig.ident", "nFeature_ADT", "nCount_ADT")}. Only for:\strong{Batch QC (Covariate Impact)}.}

\item{celltype.by}{Column name in the metadata indicating cell type information. Default is "ScType". Only for: \strong{Sample QC (Outlier sample detection by cell type), Batch QC (Clustering QC)}.}

\item{tissue}{Tissue name. Default is NULL. Only for: \strong{Sample QC (Sample quality assessment, Outlier sample detection by cell type)}.}
}
\value{
A HTML report is generated and saved to the specified output file path. No value is returned.
}
\description{
This function generates a comprehensive HTML report for single-cell data quality control (QC).
The report includes modules for sample QC, cell QC, feature QC, and batch QC, with customizable sections for each module.
}
\examples{
\dontrun{
# Preprocess data using RunPreprocess
seurat_object <- RunPreprocess(data = seurat_object, sample.by = "orig.ident")

# Generate QC report
RunReport(object = seurat_object, sample.by = "orig.ident")
}
}
